split AI file
