+++ src/ai/ai.cpp
@@ -1,4 +1,5 @@
 #include "ai.hpp"
+#include "ai_types.hpp"
 #include "system_state.hpp"
 #include "demographics.hpp"
 #include "economy_stats.hpp"
@@ -14,1014 +15,8 @@
 #include "province_templates.hpp"
 #include "triggers.hpp"
 
+
 namespace ai {
-
-enum ai_strategies {
-	industrious, militant, technological
-};
-
-float estimate_strength(sys::state& state, dcon::nation_id n) {
-	if(state.cheat_data.disable_ai) {
-		return 1.f;
-	}
-
-	float value = state.world.nation_get_military_score(n) * state.defines.alice_ai_strength_estimation_military_industrial_balance;
-	value += state.world.nation_get_industrial_score(n) * (1.f - state.defines.alice_ai_strength_estimation_military_industrial_balance);
-	for(auto subj : state.world.nation_get_overlord_as_ruler(n)) {
-		value += subj.get_subject().get_military_score() * state.defines.alice_ai_strength_estimation_military_industrial_balance;
-		value += subj.get_subject().get_industrial_score() * (1.f - state.defines.alice_ai_strength_estimation_military_industrial_balance);
-	}
-	return value;
-}
-
-float estimate_defensive_strength(sys::state& state, dcon::nation_id n) {
-	if(state.cheat_data.disable_ai) {
-		return 1.0f;
-	}
-
-	float value = estimate_strength(state, n);
-	for(auto dr : state.world.nation_get_diplomatic_relation(n)) {
-		if(!dr.get_are_allied())
-			continue;
-
-		auto other = dr.get_related_nations(0) != n ? dr.get_related_nations(0) : dr.get_related_nations(1);
-		if(other.get_overlord_as_subject().get_ruler() != n)
-			value += estimate_strength(state, other);
-	}
-	if(auto sl = state.world.nation_get_in_sphere_of(n); sl)
-		value += estimate_strength(state, sl);
-	return value;
-}
-
-float estimate_additional_offensive_strength(sys::state& state, dcon::nation_id n, dcon::nation_id target) {
-	float value = 0.f;
-	for(auto dr : state.world.nation_get_diplomatic_relation(n)) {
-		if(!dr.get_are_allied())
-			continue;
-
-		auto other = dr.get_related_nations(0) != n ? dr.get_related_nations(0) : dr.get_related_nations(1);
-		if(other.get_overlord_as_subject().get_ruler() != n && military::can_use_cb_against(state, other, target) && !military::has_truce_with(state, other, target))
-			value += estimate_strength(state, other);
-	}
-	return value * state.defines.alice_ai_offensive_strength_overestimate;
-}
-
-/* Update AI threats and rivals */
-void update_ai_general_status(sys::state& state) {
-	for(auto n : state.world.in_nation) {
-		if(state.world.nation_get_owned_province_count(n) == 0) {
-			state.world.nation_set_ai_is_threatened(n, false);
-			state.world.nation_set_ai_rival(n, dcon::nation_id{});
-			continue;
-		}
-
-		auto ll = state.world.nation_get_last_war_loss(n);
-		float safety_factor = 1.2f;
-		if(ll && state.current_date < ll + 365 * 4) {
-			safety_factor = 1.8f;
-		}
-		auto in_sphere_of = state.world.nation_get_in_sphere_of(n);
-
-		float greatest_neighbor = 0.0f;
-		for(auto b : state.world.nation_get_nation_adjacency_as_connected_nations(n)) {
-			auto other = b.get_connected_nations(0) != n ? b.get_connected_nations(0) : b.get_connected_nations(1);
-			if(!nations::are_allied(state, n, other) && (!in_sphere_of || in_sphere_of != other.get_in_sphere_of())) {
-				greatest_neighbor = std::max(greatest_neighbor, estimate_strength(state, other));
-			}
-		}
-
-		float self_str = float(state.world.nation_get_military_score(n));
-		for(auto subj : n.get_overlord_as_ruler())
-			self_str += 0.75f * float(subj.get_subject().get_military_score());
-		float defensive_str = estimate_defensive_strength(state, n);
-
-		bool threatened = defensive_str < safety_factor * greatest_neighbor;
-		state.world.nation_set_ai_is_threatened(n, threatened);
-
-		if(!n.get_ai_rival()) {
-			float min_str = 0.0f;
-			dcon::nation_id potential;
-			for(auto adj : n.get_nation_adjacency()) {
-				auto other = adj.get_connected_nations(0) != n ? adj.get_connected_nations(0) : adj.get_connected_nations(1);
-				auto ol = other.get_overlord_as_subject().get_ruler();
-				if(!ol && other.get_in_sphere_of() != n && (!threatened || !nations::are_allied(state, n, other))) {
-					auto other_str = estimate_strength(state, other);
-					if(self_str * 0.5f < other_str && other_str <= self_str * 1.5f && min_str > self_str) {
-						min_str = other_str;
-						potential = other;
-					}
-				}
-			}
-
-			if(potential) {
-				if(!n.get_is_player_controlled() && nations::are_allied(state, n, potential)) {
-					assert(command::can_cancel_alliance(state, n, potential));
-					command::execute_cancel_alliance(state, n, potential);
-				}
-				n.set_ai_rival(potential);
-			}
-		} else {
-			auto rival_str = estimate_strength(state, n.get_ai_rival());
-			auto ol = n.get_ai_rival().get_overlord_as_subject().get_ruler();
-			if(ol || n.get_ai_rival().get_in_sphere_of() == n || rival_str * 2 < self_str || self_str * 2 < rival_str) {
-				n.set_ai_rival(dcon::nation_id{});
-			}
-		}
-	}
-}
-
-// Does nation N have supremacy over target nation
-bool naval_supremacy(sys::state& state, dcon::nation_id n, dcon::nation_id target) {
-	auto self_sup = state.world.nation_get_used_naval_supply_points(n);
-
-	auto real_target = state.world.overlord_get_ruler(state.world.nation_get_overlord_as_subject(target));
-	if(!real_target)
-		real_target = target;
-
-	if(self_sup <= state.world.nation_get_used_naval_supply_points(real_target))
-		return false;
-
-	if(self_sup <= state.world.nation_get_in_sphere_of(real_target).get_used_naval_supply_points())
-		return false;
-
-	for(auto a : state.world.nation_get_diplomatic_relation(real_target)) {
-		if(!a.get_are_allied())
-			continue;
-		auto other = a.get_related_nations(0) != real_target ? a.get_related_nations(0) : a.get_related_nations(1);
-		if(self_sup <= other.get_used_naval_supply_points())
-			return false;
-	}
-
-	return true;
-}
-
-static void internal_get_alliance_targets_by_adjacency(sys::state& state, dcon::nation_id n, dcon::nation_id adj, std::vector<dcon::nation_id>& alliance_targets) {
-	for(auto nb : state.world.nation_get_nation_adjacency(adj)) {
-		auto other = nb.get_connected_nations(0) != adj ? nb.get_connected_nations(0) : nb.get_connected_nations(1);
-
-		auto our_strength = estimate_strength(state, n);
-		auto other_strength = estimate_strength(state, other);
-
-		// We really shouldn't ally very weak neighbours
-		// Surrounding ourselves with useless alliances stops expansion
-		if(other_strength <= our_strength * 0.2f) {
-			continue;
-		}
-
-		bool b = other.get_is_player_controlled()
-			? state.world.unilateral_relationship_get_interested_in_alliance(state.world.get_unilateral_relationship_by_unilateral_pair(n, other))
-			: ai_will_accept_alliance(state, other, n);
-		if(other != n && !(other.get_overlord_as_subject().get_ruler()) && !nations::are_allied(state, n, other) && !military::are_at_war(state, other, n) && b) {
-			alliance_targets.push_back(other.id);
-		}
-	}
-}
-static void internal_get_alliance_targets(sys::state& state, dcon::nation_id n, std::vector<dcon::nation_id>& alliance_targets) {
-	// Adjacency with us
-	internal_get_alliance_targets_by_adjacency(state, n, n, alliance_targets);
-	if(!alliance_targets.empty())
-		return;
-
-	// Adjacency with rival (useful for e.x, Chile allying Paraguay to fight bolivia)
-	if(auto rival = state.world.nation_get_ai_rival(n); bool(rival)) {
-		internal_get_alliance_targets_by_adjacency(state, n, rival, alliance_targets);
-		if(!alliance_targets.empty())
-			return;
-	}
-
-	// Adjacency with people who are at war with us
-	for(auto wp : state.world.nation_get_war_participant(n)) {
-		for(auto p : state.world.war_get_war_participant(wp.get_war())) {
-			if(p.get_is_attacker() == !wp.get_is_attacker()) {
-				internal_get_alliance_targets_by_adjacency(state, n, p.get_nation(), alliance_targets);
-				if(!alliance_targets.empty())
-					return;
-			}
-		}
-	}
-}
-
-void form_alliances(sys::state& state) {
-	static std::vector<dcon::nation_id> alliance_targets;
-	for(auto n : state.world.in_nation) {
-		if(!n.get_is_player_controlled() && n.get_ai_is_threatened() && !(n.get_overlord_as_subject().get_ruler())) {
-			alliance_targets.clear();
-			internal_get_alliance_targets(state, n, alliance_targets);
-			if(!alliance_targets.empty()) {
-				std::sort(alliance_targets.begin(), alliance_targets.end(), [&](dcon::nation_id a, dcon::nation_id b) {
-					if(estimate_strength(state, a) != estimate_strength(state, b))
-						return estimate_strength(state, a) > estimate_strength(state, b);
-					else
-						return a.index() > b.index();
-				});
-				if(state.world.nation_get_is_player_controlled(alliance_targets[0])) {
-					notification::post(state, notification::message{
-						[source = n](sys::state& state, text::layout_base& contents) {
-							text::add_line(state, contents, "msg_entered_automatic_alliance_1", text::variable_type::x, source);
-						},
-						"msg_entered_automatic_alliance_title",
-						n, dcon::nation_id{}, dcon::nation_id{},
-						sys::message_base_type::entered_automatic_alliance
-					});
-				}
-				nations::make_alliance(state, n, alliance_targets[0]);
-			}
-		}
-	}
-}
-
-void prune_alliances(sys::state& state) {
-	static std::vector<dcon::nation_id> prune_targets;
-	for(auto n : state.world.in_nation) {
-		if(!n.get_is_player_controlled()
-		&& !n.get_ai_is_threatened()
-		&& !(n.get_overlord_as_subject().get_ruler())) {
-			prune_targets.clear();
-			for(auto dr : n.get_diplomatic_relation()) {
-				if(dr.get_are_allied()) {
-					auto other = dr.get_related_nations(0) != n ? dr.get_related_nations(0) : dr.get_related_nations(1);
-					if(other.get_in_sphere_of() != n
-					&& !military::are_allied_in_war(state, n, other)) {
-						prune_targets.push_back(other);
-					}
-				}
-			}
-
-			if(prune_targets.empty())
-				continue;
-
-			std::sort(prune_targets.begin(), prune_targets.end(), [&](dcon::nation_id a, dcon::nation_id b) {
-				if(estimate_strength(state, a) != estimate_strength(state, b))
-					return estimate_strength(state, a) < estimate_strength(state, b);
-				else
-					return a.index() > b.index();
-			});
-
-			float greatest_neighbor = 0.0f;
-			auto in_sphere_of = state.world.nation_get_in_sphere_of(n);
-
-			for(auto b : state.world.nation_get_nation_adjacency_as_connected_nations(n)) {
-				auto other = b.get_connected_nations(0) != n ? b.get_connected_nations(0) : b.get_connected_nations(1);
-				if(!nations::are_allied(state, n, other) && (!in_sphere_of || in_sphere_of != other.get_in_sphere_of())) {
-					greatest_neighbor = std::max(greatest_neighbor, estimate_strength(state, other));
-				}
-			}
-
-			float defensive_str = estimate_defensive_strength(state, n);
-			auto ll = state.world.nation_get_last_war_loss(n);
-			float safety_factor = 1.2f;
-			if(ll && state.current_date < ll + 365 * 4) {
-				safety_factor = 1.8f;
-			}
-
-			auto safety_margin = defensive_str - safety_factor * greatest_neighbor;
-
-			for(auto pt : prune_targets) {
-				auto weakest_str = estimate_strength(state, pt);
-				if(weakest_str * 1.25 < safety_margin) {
-					safety_margin -= weakest_str;
-					assert(command::can_cancel_alliance(state, n, pt, true));
-					command::execute_cancel_alliance(state, n, pt);
-				} else if(state.world.nation_get_infamy(pt) >= state.defines.badboy_limit) {
-					safety_margin -= weakest_str;
-					assert(command::can_cancel_alliance(state, n, pt, true));
-					command::execute_cancel_alliance(state, n, pt);
-				} else {
-					break;
-				}
-			}
-		}
-	}
-}
-
-bool ai_is_close_enough(sys::state& state, dcon::nation_id target, dcon::nation_id from) {
-	auto target_continent = state.world.province_get_continent(state.world.nation_get_capital(target));
-	auto source_continent = state.world.province_get_continent(state.world.nation_get_capital(from));
-	return (target_continent == source_continent) || bool(state.world.get_nation_adjacency_by_nation_adjacency_pair(target, from));
-}
-
-static bool ai_has_mutual_enemy(sys::state& state, dcon::nation_id from, dcon::nation_id target) {
-	auto rival_a = state.world.nation_get_ai_rival(target);
-	auto rival_b = state.world.nation_get_ai_rival(from);
-	// Same rival equates to instantaneous alliance (we benefit from more allies against a common enemy)
-	if(rival_a && rival_a == rival_b)
-		return true;
-	// // Our rivals are allied?
-	// if(rival_a && rival_b && rival_a != rival_b && nations::are_allied(state, rival_a, rival_b))
-	// 	return true;
-
-	// // One of the allies of our rivals can be declared on?
-	// for(auto n : state.world.in_nation) {
-	// 	if(n.id != target && n.id != from && n.id != rival_a && n.id != rival_b) {
-	// 		if(nations::are_allied(state, rival_a, n.id) || nations::are_allied(state, rival_b, n.id)) {
-	// 			bool enemy_a = military::can_use_cb_against(state, from, n.id);
-	// 			bool enemy_b = military::can_use_cb_against(state, target, n.id);
-	// 			if(enemy_a || enemy_b)
-	// 				return true;
-	// 		}
-	// 	}
-	// }
-	return false;
-}
-
-constexpr inline float ally_overestimate = 2.f;
-
-bool ai_will_accept_alliance(sys::state& state, dcon::nation_id target, dcon::nation_id from) {
-	if(!state.world.nation_get_ai_is_threatened(target))
-		return false;
-
-	// Has not surpassed infamy limit
-	if(state.world.nation_get_infamy(target) >= state.defines.badboy_limit * 0.75f)
-		return false;
-
-	// Won't ally our rivals
-	if(state.world.nation_get_ai_rival(target) == from || state.world.nation_get_ai_rival(from) == target)
-		return false;
-
-	// No more than 2 GP allies
-	// No more than 4 alliances
-	if(bool(state.defines.alice_artificial_gp_limitant) && state.world.nation_get_is_great_power(target)) {
-		int32_t gp_count = 0;
-		for(const auto rel : state.world.nation_get_diplomatic_relation(from)) {
-			auto n = rel.get_related_nations(rel.get_related_nations(0) == from ? 1 : 0);
-			if(rel.get_are_allied() && n.get_is_great_power()) {
-				if(gp_count >= 2) {
-					return false;
-				}
-				++gp_count;
-			}
-		}
-	}
-	if(bool(state.defines.alice_spherelings_only_ally_sphere)) {
-		auto spherelord = state.world.nation_get_in_sphere_of(from);
-		//If no spherelord -> Then must not ally spherelings
-		//If spherelord -> Then must not ally non-spherelings
-		if(state.world.nation_get_in_sphere_of(target) != spherelord && target != spherelord)
-			return false;
-		if(target == spherelord)
-			return true; //always ally spherelord
-	}
-
-	auto natid = state.world.nation_get_identity_from_identity_holder(from);
-	for(auto prov_owner : state.world.nation_get_province_ownership(target)) {
-		auto prov = prov_owner.get_province();
-
-		for(auto core : prov.get_core_as_province()) {
-			if(core.get_identity() == natid) {
-				return false; // holds our cores
-			}
-		}
-	}
-
-	if(ai_has_mutual_enemy(state, from, target))
-		return true;
-
-	// Otherwise we may consider alliances only iff they are close to our continent or we are adjacent
-	if(!ai_is_close_enough(state, target, from))
-		return false;
-
-	// And also if they're powerful enough to be considered for an alliance
-	auto target_score = estimate_strength(state, target);
-	auto source_score = estimate_strength(state, from);
-	return std::max<float>(source_score, 1.f) * ally_overestimate >= target_score;
-}
-
-bool ai_will_accept_free_trade(sys::state& state, dcon::nation_id target, dcon::nation_id from) {
-	return false;
-}
-
-void explain_ai_trade_agreement_reasons(sys::state& state, dcon::nation_id target, text::layout_base& contents, int32_t indent) {
-	text::add_line_with_condition(state, contents, "never", false, indent + 15);
-}
-
-void explain_ai_alliance_reasons(sys::state& state, dcon::nation_id target, text::layout_base& contents, int32_t indent) {
-
-	text::add_line_with_condition(state, contents, "ai_alliance_1", state.world.nation_get_ai_is_threatened(target), indent);
-
-	text::add_line(state, contents, "kierkegaard_1", indent);
-
-	text::add_line_with_condition(state, contents, "ai_alliance_5", ai_has_mutual_enemy(state, state.local_player_nation, target), indent + 15);
-
-	text::add_line(state, contents, "kierkegaard_2", indent);
-
-	text::add_line_with_condition(state, contents, "ai_alliance_2", ai_is_close_enough(state, target, state.local_player_nation), indent + 15);
-
-	text::add_line_with_condition(state, contents, "ai_alliance_3", state.world.nation_get_ai_rival(target) != state.local_player_nation && state.world.nation_get_ai_rival(state.local_player_nation) != target, indent + 15);
-
-	auto target_score = estimate_strength(state, target);
-	auto source_score = estimate_strength(state, state.local_player_nation);
-	text::add_line_with_condition(state, contents, "ai_alliance_4", std::max<float>(source_score, 1.f) * ally_overestimate >= target_score, indent + 15);
-
-	auto holdscores = false;
-	auto natid = state.world.nation_get_identity_from_identity_holder(state.local_player_nation);
-	for(auto prov_owner : state.world.nation_get_province_ownership(target)) {
-		auto prov = prov_owner.get_province();
-
-		for(auto core : prov.get_core_as_province()) {
-			if(core.get_identity() == natid) {
-				holdscores = true; // holds our cores
-			}
-		}
-	}
-
-	text::add_line_with_condition(state, contents, "ai_alliance_5", !holdscores, indent + 15);
-}
-
-bool ai_will_grant_access(sys::state& state, dcon::nation_id target, dcon::nation_id from) {
-	if(!state.world.nation_get_is_at_war(from))
-		return false;
-	if(state.world.nation_get_ai_rival(target) == from)
-		return false;
-	if(military::are_at_war(state, from, state.world.nation_get_ai_rival(target)))
-		return true;
-
-	for(auto wa : state.world.nation_get_war_participant(target)) {
-		auto is_attacker = wa.get_is_attacker();
-		for(auto o : wa.get_war().get_war_participant()) {
-			if(o.get_is_attacker() != is_attacker) {
-				if(military::are_at_war(state, o.get_nation(), from))
-					return true;
-			}
-		}
-	}
-	return false;
-
-}
-void explain_ai_access_reasons(sys::state& state, dcon::nation_id target, text::layout_base& contents, int32_t indent) {
-	text::add_line_with_condition(state, contents, "ai_access_1", ai_will_grant_access(state, target, state.local_player_nation), indent);
-}
-
-// MP compliant
-void update_ai_research(sys::state& state) {
-	auto ymd_date = state.current_date.to_ymd(state.start_date);
-	auto year = uint32_t(ymd_date.year);
-	concurrency::parallel_for(uint32_t(0), state.world.nation_size(), [&](uint32_t id) {
-		dcon::nation_id n{ dcon::nation_id::value_base_t(id) };
-
-		if(state.world.nation_get_is_player_controlled(n)
-			|| state.world.nation_get_current_research(n)
-			|| !state.world.nation_get_is_civilized(n)
-			|| state.world.nation_get_owned_province_count(n) == 0) {
-
-			//skip -- does not need new research
-			return;
-		}
-
-		struct potential_techs {
-			dcon::technology_id id;
-			float weight = 0.0f;
-		};
-
-		std::vector<potential_techs> potential;
-
-		for(auto tid : state.world.in_technology) {
-			if(state.world.nation_get_active_technologies(n, tid))
-				continue; // Already researched
-
-			if(state.current_date.to_ymd(state.start_date).year >= state.world.technology_get_year(tid)) {
-				// Research technologies costing LP (military doctrines) only if can research it immediately
-				if(culture::effective_technology_lp_cost(state, year, n, tid) > state.world.nation_get_leadership_points(n)) {
-					continue;
-				}
-				// Technologies costing RP:
-				// Find previous technology before this one
-				dcon::technology_id prev_tech = dcon::technology_id(dcon::technology_id::value_base_t(tid.id.index() - 1));
-				// Previous technology is from the same folder so we have to check that we have researched it beforehand
-				if(tid.id.index() != 0 && state.world.technology_get_folder_index(prev_tech) == state.world.technology_get_folder_index(tid)) {
-					// Only allow if all previously researched techs are researched
-					if(state.world.nation_get_active_technologies(n, prev_tech))
-						potential.push_back(potential_techs{ tid, 0.0f });
-				} else { // first tech in folder
-					potential.push_back(potential_techs{ tid, 0.0f });
-				}
-			}
-		}
-
-		for(auto& pt : potential) { // weight techs
-			auto base = state.world.technology_get_ai_weight(pt.id);
-			if(state.world.nation_get_ai_is_threatened(n) && state.culture_definitions.tech_folders[state.world.technology_get_folder_index(pt.id)].category == culture::tech_category::army) {
-				base *= 2.0f;
-			}
-
-			if(state.world.nation_get_ai_strategy(n) == ai_strategies::militant && state.culture_definitions.tech_folders[state.world.technology_get_folder_index(pt.id)].category == culture::tech_category::army) {
-				base *= 2.0f;
-			}
-			if(state.world.nation_get_ai_strategy(n) == ai_strategies::industrious && state.culture_definitions.tech_folders[state.world.technology_get_folder_index(pt.id)].category == culture::tech_category::industry) {
-				base *= 2.0f;
-			}
-			if(state.world.nation_get_ai_strategy(n) == ai_strategies::industrious && state.culture_definitions.tech_folders[state.world.technology_get_folder_index(pt.id)].category == culture::tech_category::commerce) {
-				base *= 2.0f;
-			}
-			if(state.world.nation_get_ai_strategy(n) == ai_strategies::technological && state.culture_definitions.tech_folders[state.world.technology_get_folder_index(pt.id)].category == culture::tech_category::culture) {
-				base *= 2.0f;
-			}
-
-			auto cost = std::max(1.0f, culture::effective_technology_rp_cost(state, year, n, pt.id));
-			pt.weight = base / cost;
-		}
-		auto rval = rng::get_random(state, id);
-		std::sort(potential.begin(), potential.end(), [&](potential_techs& a, potential_techs& b) {
-			if(a.weight != b.weight)
-				return a.weight > b.weight;
-			else // sort semi randomly
-				return (a.id.index() ^ rval) > (b.id.index() ^ rval);
-		});
-
-		if(!potential.empty()) {
-			state.world.nation_set_current_research(n, potential[0].id);
-		}
-	});
-}
-
-void initialize_ai_tech_weights(sys::state& state) {
-	for(auto t : state.world.in_technology) {
-		float base = 1000.0f;
-		if(state.culture_definitions.tech_folders[t.get_folder_index()].category == culture::tech_category::army)
-			base *= 3.0f;
-
-		if(t.get_increase_building(economy::province_building_type::naval_base))
-			base *= 4.5f;
-		else if(state.culture_definitions.tech_folders[t.get_folder_index()].category == culture::tech_category::navy)
-			base *= 0.9f;
-
-		auto mod = t.get_modifier();
-		auto& vals = mod.get_national_values();
-		for(uint32_t i = 0; i < sys::national_modifier_definition::modifier_definition_size; ++i) {
-			if(vals.offsets[i] == sys::national_mod_offsets::research_points) {
-				base *= 6.0f;
-			} else if(vals.offsets[i] == sys::national_mod_offsets::research_points_modifier) {
-				base *= 6.0f;
-			} else if(vals.offsets[i] == sys::national_mod_offsets::education_efficiency) {
-				base *= 3.0f;
-			} else if(vals.offsets[i] == sys::national_mod_offsets::education_efficiency_modifier) {
-				base *= 3.0f;
-			} else if(vals.offsets[i] == sys::national_mod_offsets::pop_growth) {
-				base *= 5.0f;
-			} else if(vals.offsets[i] == sys::national_mod_offsets::max_national_focus) {
-				base *= 4.0f;
-			} else if(vals.offsets[i] == sys::national_mod_offsets::colonial_life_rating) {
-				base *= 4.0f;
-			} else if(vals.offsets[i] == sys::national_mod_offsets::rgo_output) {
-				base *= 1.2f;
-			} else if(vals.offsets[i] == sys::national_mod_offsets::factory_output) {
-				base *= 3.0f;
-			} else if(vals.offsets[i] == sys::national_mod_offsets::factory_throughput) {
-				base *= 5.0f;
-			} else if(vals.offsets[i] == sys::national_mod_offsets::factory_input) {
-				base *= 3.0f;
-			} else if(vals.offsets[i] == sys::national_mod_offsets::tax_efficiency) {
-				base *= 3.0f;
-			}
-		}
-
-		t.set_ai_weight(base);
-	}
-}
-
-void update_factory_types_priority(sys::state& state) {
-	concurrency::parallel_for(uint32_t(0), state.world.nation_size(), [&](uint32_t id) {
-		dcon::nation_id n{ dcon::nation_id::value_base_t(id) };
-
-		if(state.world.nation_get_owned_province_count(n) == 0) {
-			// skip -- do not need experience
-			return;
-		}
-		// private research
-
-		std::vector<float> employment;
-		employment.resize(state.world.factory_type_size());
-
-		state.world.nation_for_each_province_ownership_as_nation(n, [&](dcon::province_ownership_id province_ownership) {
-			auto province = state.world.province_ownership_get_province(province_ownership);
-			auto sid = state.world.province_get_state_membership(province);
-			auto market = state.world.state_instance_get_market_from_local_market(sid);
-			state.world.province_for_each_factory_location_as_province(province, [&](auto location) {
-				auto factory = state.world.factory_location_get_factory(location);
-				auto type = state.world.factory_get_building_type(factory);
-				employment[type.id.index()] +=
-					state.world.factory_get_primary_employment(factory)
-					* economy::factory_type_input_cost(
-						state, n, market, type
-					);
-			});
-		});
-
-		state.world.for_each_factory_type([&](dcon::factory_type_id factory_type_id) {
-
-			if(
-				!state.world.nation_get_active_building(n, factory_type_id)
-				&& !state.world.factory_type_get_is_available_from_start(factory_type_id)
-			) {
-				state.world.nation_set_factory_type_experience_priority_private(n, factory_type_id, 0.f);
-				return;
-			}
-
-			// capitalists want to minimise costs related to current factories
-			// so they choose factory types
-			// according to:
-			// 1) current amount of workers in factories of a given type
-			// 2) costs of inputs
-
-			auto costs = employment[factory_type_id.index()];
-			state.world.nation_set_factory_type_experience_priority_private(n, factory_type_id, costs);
-		});
-
-		if(state.world.nation_get_is_player_controlled(n)) {
-			// skip -- do not need AI
-			return;
-		}
-		// national research
-
-		std::vector<float> supply;
-		supply.resize(state.world.commodity_size());
-
-		state.world.for_each_commodity([&](dcon::commodity_id cid) {
-			supply[cid.index()] = economy::supply(state, n, cid);
-		});
-
-		state.world.for_each_factory_type([&](dcon::factory_type_id factory_type_id) {
-			if(
-				!state.world.nation_get_active_building(n, factory_type_id)
-				&& !state.world.factory_type_get_is_available_from_start(factory_type_id)
-			) {
-				state.world.nation_set_factory_type_experience_priority_private(n, factory_type_id, 0.f);
-				return;
-			}
-
-			// auto priority = 0.5f + float(rng::get_random(state, uint32_t(id) ^ uint32_t(factory_type_id)) & 0xFFFF) / float(0xFFFF);
-
-			auto& inputs = state.world.factory_type_get_inputs(factory_type_id);
-
-			auto min_effective_supply = std::numeric_limits<float>::infinity();
-			for(uint32_t i = 0; i < economy::commodity_set::set_size; ++i) {
-				auto input = inputs.commodity_type[i];
-				if(input) {
-					min_effective_supply = std::min(supply[input.index()] / inputs.commodity_amounts[i], min_effective_supply);
-				} else {
-					break;
-				}
-			}
-
-			//check if there are "rivals" which would push you away from the industry
-			auto local_price = economy::price(state, n, state.world.factory_type_get_output(factory_type_id));
-			auto rival_price = local_price * 2.f;
-			for(auto adj : state.world.nation_get_nation_adjacency(n)) {
-				auto other = adj.get_connected_nations(0) != n ? adj.get_connected_nations(0) : adj.get_connected_nations(1);
-				rival_price = std::min(rival_price, economy::price(state, other, state.world.factory_type_get_output(factory_type_id)));
-			}
-
-			auto rival_modifier = (rival_price + 0.01f) / (local_price + 0.01f);
-
-			// we want this modifier to be really strong
-			rival_modifier = (rival_modifier * rival_modifier) * (rival_modifier * rival_modifier) * rival_modifier;
-
-			state.world.nation_set_factory_type_experience_priority_national(n, factory_type_id, min_effective_supply * rival_modifier);
-		});
-	});
-}
-
-void update_influence_priorities(sys::state& state) {
-	struct weighted_nation {
-		dcon::nation_id id;
-		float weight = 0.0f;
-	};
-	static std::vector<weighted_nation> targets;
-
-	for(auto gprl : state.world.in_gp_relationship) {
-		if(gprl.get_great_power().get_is_player_controlled()) {
-			// nothing -- player GP
-		} else {
-			auto& status = gprl.get_status();
-			status &= ~nations::influence::priority_mask;
-			if((status & nations::influence::level_mask) == nations::influence::level_in_sphere) {
-				status |= nations::influence::priority_one;
-			}
-		}
-	}
-
-	for(auto& n : state.great_nations) {
-		if(state.world.nation_get_is_player_controlled(n.nation))
-			continue;
-
-		targets.clear();
-		for(auto t : state.world.in_nation) {
-			if(t.get_is_great_power())
-				continue;
-			if(t.get_owned_province_count() == 0)
-				continue;
-			if(t.get_in_sphere_of() == n.nation)
-				continue;
-			if(t.get_demographics(demographics::total) > state.defines.large_population_limit)
-				continue;
-
-			auto natid = state.world.nation_get_identity_from_identity_holder(n.nation);
-			auto holdscores = false;
-			for(auto prov_owner : state.world.nation_get_province_ownership(t)) {
-				auto prov = prov_owner.get_province();
-
-				for(auto core : prov.get_core_as_province()) {
-					if(core.get_identity() == natid) {
-						holdscores = true;
-						break;
-					}
-				}
-
-				if(holdscores)
-					break;
-			}
-
-			if(holdscores) {
-				continue; // holds our cores
-			}
-
-			float weight = 0.0f;
-
-			for(auto c : state.world.in_commodity) {
-				if(auto d = economy::demand(state, n.nation, c); d > 0.001f) {
-					auto cweight = std::min(
-						1.0f,
-						economy::supply(state, n.nation, c) / d)
-						* (1.0f - economy::demand_satisfaction(state, n.nation, c));
-					weight += cweight;
-				}
-			}
-
-			//We probably don't want to fight a forever lasting sphere war, let's find some other uncontested nations
-			if(t.get_in_sphere_of()) {
-				weight /= 4.0f;
-			}
-
-			//Prioritize primary culture before culture groups; should ensure Prussia spheres all of the NGF first before trying to contest Austria
-			if(t.get_primary_culture() == state.world.nation_get_primary_culture(n.nation)) {
-				weight += 1.0f;
-				weight *= 4000.0f;
-			}
-
-			else if(t.get_primary_culture().get_group_from_culture_group_membership() == state.world.nation_get_primary_culture(n.nation).get_group_from_culture_group_membership()) {
-				weight *= 4.0f;
-			}
-			//Focus on gaining influence against nations we have active wargoals against so we can remove their protector, even if it's us
-			if(military::can_use_cb_against(state, n.nation, t) && t.get_in_sphere_of()) {
-				weight += 1.0f;
-				weight *= 1000.0f;
-			}
-			//If it doesn't neighbor us or a friendly sphere and isn't coastal, please don't sphere it, we don't want sphere gore
-			bool is_reachable = false;
-			for(auto adj : state.world.nation_get_nation_adjacency(t)) {
-				auto casted_adj = adj.get_connected_nations(0) != t ? adj.get_connected_nations(0) : adj.get_connected_nations(1);
-				if(casted_adj == n.nation) {
-					is_reachable = true;
-					break;
-				}
-				if(casted_adj.get_in_sphere_of() == n.nation) {
-					is_reachable = true;
-					break;
-				}
-			};
-
-			//Is coastal? Technically reachable
-			if(state.world.nation_get_central_ports(t) > 0) {
-				is_reachable = true;
-			}
-
-			//Prefer neighbors
-			if(state.world.get_nation_adjacency_by_nation_adjacency_pair(n.nation, t.id)) {
-				weight *= 10.0f;
-				is_reachable = true;
-			}
-
-			if(!is_reachable) {
-				weight *= 0.0f;
-			}
-
-			targets.push_back(weighted_nation{ t.id, weight });
-		}
-
-		std::sort(targets.begin(), targets.end(), [](weighted_nation const& a, weighted_nation const& b) {
-			if(a.weight != b.weight)
-				return a.weight > b.weight;
-			else
-				return a.id.index() < b.id.index();
-		});
-
-		uint32_t i = 0;
-		for(; i < 2 && i < targets.size(); ++i) {
-			auto rel = state.world.get_gp_relationship_by_gp_influence_pair(targets[i].id, n.nation);
-			if(!rel)
-				rel = state.world.force_create_gp_relationship(targets[i].id, n.nation);
-			state.world.gp_relationship_get_status(rel) |= nations::influence::priority_three;
-		}
-		for(; i < 4 && i < targets.size(); ++i) {
-			auto rel = state.world.get_gp_relationship_by_gp_influence_pair(targets[i].id, n.nation);
-			if(!rel)
-				rel = state.world.force_create_gp_relationship(targets[i].id, n.nation);
-			state.world.gp_relationship_get_status(rel) |= nations::influence::priority_two;
-		}
-		for(; i < 6 && i < targets.size(); ++i) {
-			auto rel = state.world.get_gp_relationship_by_gp_influence_pair(targets[i].id, n.nation);
-			if(!rel)
-				rel = state.world.force_create_gp_relationship(targets[i].id, n.nation);
-			state.world.gp_relationship_get_status(rel) |= nations::influence::priority_one;
-		}
-	}
-}
-
-void perform_influence_actions(sys::state& state) {
-	for(auto gprl : state.world.in_gp_relationship) {
-		if(gprl.get_great_power().get_is_player_controlled()) {
-			// nothing -- player GP
-		} else {
-			if((gprl.get_status() & nations::influence::is_banned) != 0)
-				continue; // can't do anything with a banned nation
-
-			if(military::are_at_war(state, gprl.get_great_power(), gprl.get_influence_target()))
-				continue; // can't do anything while at war
-
-			auto clevel = (nations::influence::level_mask & gprl.get_status());
-			if(clevel == nations::influence::level_in_sphere)
-				continue; // already in sphere
-
-			auto current_sphere = gprl.get_influence_target().get_in_sphere_of();
-
-			if(state.defines.increaseopinion_influence_cost <= gprl.get_influence() && clevel != nations::influence::level_friendly) {
-				assert(command::can_increase_opinion(state, gprl.get_great_power(), gprl.get_influence_target()));
-				command::execute_increase_opinion(state, gprl.get_great_power(), gprl.get_influence_target());
-			} else if(state.defines.removefromsphere_influence_cost <= gprl.get_influence() && current_sphere /* && current_sphere != gprl.get_great_power()*/ && clevel == nations::influence::level_friendly) { // condition taken care of by check above
-				assert(command::can_remove_from_sphere(state, gprl.get_great_power(), gprl.get_influence_target(), gprl.get_influence_target().get_in_sphere_of()));
-				command::execute_remove_from_sphere(state, gprl.get_great_power(), gprl.get_influence_target(), gprl.get_influence_target().get_in_sphere_of());
-			} else if(state.defines.addtosphere_influence_cost <= gprl.get_influence() && !current_sphere && clevel == nations::influence::level_friendly) {
-				assert(command::can_add_to_sphere(state, gprl.get_great_power(), gprl.get_influence_target()));
-				command::execute_add_to_sphere(state, gprl.get_great_power(), gprl.get_influence_target());
-			//De-sphere countries we have wargoals against, desphering countries need to check for going over infamy
-			} else if(military::can_use_cb_against(state, gprl.get_great_power(), gprl.get_influence_target())
-				&& state.defines.removefromsphere_influence_cost <= gprl.get_influence()
-				&& current_sphere
-				&& clevel == nations::influence::level_friendly
-				&& (state.world.nation_get_infamy(gprl.get_great_power()) + state.defines.removefromsphere_infamy_cost) < state.defines.badboy_limit
-			) {
-				assert(command::can_remove_from_sphere(state, gprl.get_great_power(), gprl.get_influence_target(), gprl.get_influence_target().get_in_sphere_of()));
-				command::execute_remove_from_sphere(state, gprl.get_great_power(), gprl.get_influence_target(), gprl.get_influence_target().get_in_sphere_of());
-			}
-		}
-	}
-}
-
-void identify_focuses(sys::state& state) {
-	for(auto f : state.world.in_national_focus) {
-		if(f.get_promotion_amount() > 0) {
-			if(f.get_promotion_type() == state.culture_definitions.clergy)
-				state.national_definitions.clergy_focus = f;
-			if(f.get_promotion_type() == state.culture_definitions.soldiers)
-				state.national_definitions.soldier_focus = f;
-			if(f.get_promotion_type() == state.culture_definitions.aristocrat)
-				state.national_definitions.aristocrat_focus = f;
-			if(f.get_promotion_type() == state.culture_definitions.capitalists)
-				state.national_definitions.capitalist_focus = f;
-			if(f.get_promotion_type() == state.culture_definitions.primary_factory_worker)
-				state.national_definitions.primary_factory_worker_focus = f;
-			if(f.get_promotion_type() == state.culture_definitions.secondary_factory_worker)
-				state.national_definitions.secondary_factory_worker_focus = f;
-		}
-	}
-}
-
-void update_focuses(sys::state& state) {
-	for(auto si : state.world.in_state_instance) {
-		if(!si.get_nation_from_state_ownership().get_is_player_controlled())
-			si.set_owner_focus(dcon::national_focus_id{});
-	}
-
-	for(auto n : state.world.in_nation) {
-		if(n.get_is_player_controlled())
-			continue;
-		if(n.get_owned_province_count() == 0)
-			continue;
-
-		n.set_state_from_flashpoint_focus(dcon::state_instance_id{});
-
-		auto num_focuses_total = nations::max_national_focuses(state, n);
-		if(num_focuses_total <= 0)
-			return;
-
-		auto base_opt = state.world.pop_type_get_research_optimum(state.culture_definitions.clergy);
-		auto clergy_frac = n.get_demographics(demographics::to_key(state, state.culture_definitions.clergy)) / n.get_demographics(demographics::total);
-		bool max_clergy = clergy_frac >= base_opt;
-
-		static std::vector<dcon::state_instance_id> ordered_states;
-		ordered_states.clear();
-		for(auto si : n.get_state_ownership()) {
-			ordered_states.push_back(si.get_state().id);
-		}
-		std::sort(ordered_states.begin(), ordered_states.end(), [&](auto a, auto b) {
-			auto apop = state.world.state_instance_get_demographics(a, demographics::total);
-			auto bpop = state.world.state_instance_get_demographics(b, demographics::total);
-			if(apop != bpop)
-				return apop > bpop;
-			else
-				return a.index() < b.index();
-		});
-		bool threatened = n.get_ai_is_threatened() || n.get_is_at_war();
-		for(uint32_t i = 0; num_focuses_total > 0 && i < ordered_states.size(); ++i) {
-			auto prov = state.world.state_instance_get_capital(ordered_states[i]);
-			if(max_clergy) {
-				if(threatened) {
-					auto nf = state.national_definitions.soldier_focus;
-					auto k = state.world.national_focus_get_limit(nf);
-					if(!k || trigger::evaluate(state, k, trigger::to_generic(prov), trigger::to_generic(n), -1)) {
-						assert(command::can_set_national_focus(state, n, ordered_states[i], nf));
-						state.world.state_instance_set_owner_focus(ordered_states[i], state.national_definitions.soldier_focus);
-						--num_focuses_total;
-					}
-				} else {
-					auto total = state.world.state_instance_get_demographics(ordered_states[i], demographics::total);
-					auto cfrac = state.world.state_instance_get_demographics(ordered_states[i], demographics::to_key(state, state.culture_definitions.clergy)) / total;
-					if(cfrac < state.defines.max_clergy_for_literacy * 0.8f && !state.world.province_get_is_colonial(state.world.state_instance_get_capital(ordered_states[i]))) {
-						auto nf = state.national_definitions.clergy_focus;
-						auto k = state.world.national_focus_get_limit(nf);
-						if(!k || trigger::evaluate(state, k, trigger::to_generic(prov), trigger::to_generic(n), -1)) {
-							assert(command::can_set_national_focus(state, n, ordered_states[i], nf));
-							state.world.state_instance_set_owner_focus(ordered_states[i], state.national_definitions.clergy_focus);
-							--num_focuses_total;
-						}
-					}
-				}
-			} else {
-				// If we haven't maxxed out clergy on this state, then our number 1 priority is to maximize clergy
-				auto cfrac = state.world.state_instance_get_demographics(ordered_states[i], demographics::to_key(state, state.culture_definitions.clergy)) / state.world.state_instance_get_demographics(ordered_states[i], demographics::total);
-				if(cfrac < base_opt * 1.2f && !state.world.province_get_is_colonial(state.world.state_instance_get_capital(ordered_states[i]))) {
-					auto nf = state.national_definitions.clergy_focus;
-					auto k = state.world.national_focus_get_limit(nf);
-					if(!k || trigger::evaluate(state, k, trigger::to_generic(prov), trigger::to_generic(n), -1)) {
-						assert(command::can_set_national_focus(state, n, ordered_states[i], nf));
-						state.world.state_instance_set_owner_focus(ordered_states[i], state.national_definitions.clergy_focus);
-						--num_focuses_total;
-					}
-				}
-			}
-		}
-
-		for(uint32_t i = 0; num_focuses_total > 0 && i < ordered_states.size(); ++i) {
-			auto prov = state.world.state_instance_get_capital(ordered_states[i]);
-
-			if(state.world.province_get_is_colonial(state.world.state_instance_get_capital(ordered_states[i]))) {
-				continue;
-			}
-
-			auto total = state.world.state_instance_get_demographics(ordered_states[i], demographics::total);
-			auto pw_num = state.world.state_instance_get_demographics(ordered_states[i], demographics::to_key(state, state.culture_definitions.primary_factory_worker));
-			auto pw_employed = state.world.state_instance_get_demographics(ordered_states[i], demographics::to_employment_key(state, state.culture_definitions.primary_factory_worker));
-			auto sw_num = state.world.state_instance_get_demographics(ordered_states[i], demographics::to_key(state, state.culture_definitions.secondary_factory_worker));
-			auto sw_employed = state.world.state_instance_get_demographics(ordered_states[i], demographics::to_employment_key(state, state.culture_definitions.secondary_factory_worker));
-			auto sw_frac = sw_num / std::max(pw_num + sw_num, 1.0f);
-			auto ideal_swfrac = (1.f - state.economy_definitions.craftsmen_fraction);
-			// Due to floating point comparison where 2.9999 != 3, we will round the number
-			// so that the ratio is NOT exact, but rather an aproximate
-			if((pw_employed >= pw_num || pw_num < 1.0f)) {
-				auto nf = state.national_definitions.primary_factory_worker_focus;
-				auto k = state.world.national_focus_get_limit(nf);
-				if(!k || trigger::evaluate(state, k, trigger::to_generic(prov), trigger::to_generic(n), -1)) {
-					// Keep balance between ratio of factory workers
-					// we will only promote secondary workers if none are unemployed
-					assert(command::can_set_national_focus(state, n, ordered_states[i], nf));
-					state.world.state_instance_set_owner_focus(ordered_states[i], nf);
-					--num_focuses_total;
-				}
-			} else if(pw_num > 1.0f && pw_employed > 1.0f && int8_t(sw_frac * 100.f) != int8_t(ideal_swfrac * 100.f)) {
-				auto nf = state.national_definitions.secondary_factory_worker_focus;
-				auto k = state.world.national_focus_get_limit(nf);
-				if(!k || trigger::evaluate(state, k, trigger::to_generic(prov), trigger::to_generic(n), -1)) {
-					// Keep balance between ratio of factory workers
-					// we will only promote primary workers if none are unemployed
-					assert(command::can_set_national_focus(state, n, ordered_states[i], nf));
-					state.world.state_instance_set_owner_focus(ordered_states[i], nf);
-					--num_focuses_total;
-				}
-			} else {
-				/* If we are a civilized nation, and we allow pops to operate on the economy
-				   i.e Laissez faire, we WILL promote capitalists, since they will help to
-				   build new factories for us */
-				auto rules = n.get_combined_issue_rules();
-				if(n.get_is_civilized() && (rules & (issue_rule::pop_build_factory | issue_rule::pop_build_factory_invest | issue_rule::pop_expand_factory | issue_rule::pop_expand_factory_invest | issue_rule::pop_open_factory | issue_rule::pop_open_factory_invest)) != 0) {
-					auto nf = state.national_definitions.capitalist_focus;
-					auto k = state.world.national_focus_get_limit(nf);
-					if(!k || trigger::evaluate(state, k, trigger::to_generic(prov), trigger::to_generic(n), -1)) {
-						assert(command::can_set_national_focus(state, n, ordered_states[i], nf));
-						state.world.state_instance_set_owner_focus(ordered_states[i], nf);
-						--num_focuses_total;
-					}
-				} else {
-					auto nf = state.national_definitions.aristocrat_focus;
-					auto k = state.world.national_focus_get_limit(nf);
-					if(!k || trigger::evaluate(state, k, trigger::to_generic(prov), trigger::to_generic(n), -1)) {
-						assert(command::can_set_national_focus(state, n, ordered_states[i], nf));
-						state.world.state_instance_set_owner_focus(ordered_states[i], nf);
-						--num_focuses_total;
-					}
-				}
-			}
-		}
-	}
-}
 
 void take_ai_decisions(sys::state& state) {
 	using decision_nation_pair = std::pair<dcon::decision_id, dcon::nation_id>;
@@ -1163,263 +158,6 @@ void update_ai_ruling_party(sys::state& state) {
 	}
 }
 
-void get_craved_factory_types(sys::state& state, dcon::nation_id nid, dcon::market_id mid, dcon::province_id pid, std::vector<dcon::factory_type_id>& desired_types, bool pop_project) {
-	assert(desired_types.empty());
-	auto n = dcon::fatten(state.world, nid);
-	auto m = dcon::fatten(state.world, mid);
-	auto sid = m.get_zone_from_local_market();
-
-	auto const tax_eff = economy::tax_collection_rate(state, nid, pid);
-	auto const rich_effect = (1.0f - tax_eff * float(state.world.nation_get_rich_tax(n)) / 100.0f);
-	auto wage = state.world.province_get_labor_price(pid, economy::labor::basic_education) * 2.f;
-
-	if(desired_types.empty()) {
-		for(auto type : state.world.in_factory_type) {
-			if(n.get_active_building(type) || type.get_is_available_from_start()) {
-				float cost = economy::factory_type_build_cost(state, n, pid, type, pop_project);
-				float output = economy::factory_type_output_cost(state, n, m, type);
-				float input = economy::factory_type_input_cost(state, n, m, type);
-
-				auto profit = (output - input - wage * type.get_base_workforce()) * (1.0f - rich_effect);
-				auto roi = profit / cost;
-
-				if(profit / input > 10.f && roi > 0.01f)
-					desired_types.push_back(type.id);
-			} // END if building unlocked
-		}
-	}
-}
-
-void get_desired_factory_types(sys::state& state, dcon::nation_id nid, dcon::market_id mid, dcon::province_id pid, std::vector<dcon::factory_type_id>& desired_types, bool pop_project) {
-	assert(desired_types.empty());
-	auto n = dcon::fatten(state.world, nid);
-	auto m = dcon::fatten(state.world, mid);
-	auto sid = m.get_zone_from_local_market();
-
-	auto const tax_eff = economy::tax_collection_rate(state, nid, pid);
-	auto const rich_effect = (1.0f - tax_eff * float(state.world.nation_get_rich_tax(n)) / 100.0f);
-	auto wage = state.world.province_get_labor_price(pid, economy::labor::basic_education) * 2.f;
-
-	// pass zero:
-	// factories with stupid income margins
-	// which are impossible to ignore if you are sane
-	if(desired_types.empty()) {
-		for(auto type : state.world.in_factory_type) {
-			if(n.get_active_building(type) || type.get_is_available_from_start()) {
-				auto& inputs = type.get_inputs();
-				bool lacking_input = false;
-				bool lacking_output = m.get_demand_satisfaction(type.get_output()) < 0.98f;
-
-				for(uint32_t i = 0; i < economy::commodity_set::set_size; ++i) {
-					if(inputs.commodity_type[i]) {
-						if(m.get_demand_satisfaction(inputs.commodity_type[i]) < 0.5f)
-							lacking_input = true;
-					} else {
-						break;
-					}
-				}
-
-				auto& constr_cost = type.get_construction_costs();
-				auto lacking_constr = false;
-
-				for(uint32_t i = 0; i < economy::commodity_set::set_size; ++i) {
-					if(constr_cost.commodity_type[i]) {
-						// If there is absolute deficit of construction goods, don't build for now
-						// However, we're interested in this market signal only if there are any transactions happening
-						if(m.get_demand(constr_cost.commodity_type[i]) > 0.01f && m.get_demand_satisfaction(constr_cost.commodity_type[i]) < 0.1f)
-							lacking_constr = true;
-					} else {
-						break;
-					}
-				}
-
-				float cost = economy::factory_type_build_cost(state, n, pid, type, pop_project) + 0.1f;
-				float output = economy::factory_type_output_cost(state, n, m, type);
-				float input = economy::factory_type_input_cost(state, n, m, type) + 0.1f;
-
-				auto profit = (output - input - wage * type.get_base_workforce()) * (1.0f - rich_effect);
-				auto roi = profit / cost;
-
-				if(!lacking_constr && profit / input > 2.f && roi > 0.01f)
-					desired_types.push_back(type.id);
-			} // END if building unlocked
-		}
-	}
-
-	// first pass: try to create factories which will pay back investment fast - in a year at most:
-	// or have very high margins
-	if(desired_types.empty()) { 
-		for(auto type : state.world.in_factory_type) {
-			if(n.get_active_building(type) || type.get_is_available_from_start()) {
-				auto& inputs = type.get_inputs();
-				bool lacking_input = false;
-				bool lacking_output = m.get_demand_satisfaction(type.get_output()) < 0.98f;
-
-				for(uint32_t i = 0; i < economy::commodity_set::set_size; ++i) {
-					if(inputs.commodity_type[i]) {
-						if(m.get_demand_satisfaction(inputs.commodity_type[i]) < 0.5f)
-							lacking_input = true;
-					} else {
-						break;
-					}
-				}
-
-				auto& constr_cost = type.get_construction_costs();
-				auto lacking_constr = false;
-
-				for(uint32_t i = 0; i < economy::commodity_set::set_size; ++i) {
-					if(constr_cost.commodity_type[i]) {
-						// If there is absolute deficit of construction goods, don't build for now
-						// However, we're interested in this market signal only if there are any transactions happening
-						if(m.get_demand(constr_cost.commodity_type[i]) > 0.01f && m.get_demand_satisfaction(constr_cost.commodity_type[i]) < 0.1f)
-							lacking_constr = true;
-					} else {
-						break;
-					}
-				}
-
-				float cost = economy::factory_type_build_cost(state, n, pid, type, pop_project) + 0.1f;
-				float output = economy::factory_type_output_cost(state, n, m, type);
-				float input = economy::factory_type_input_cost(state, n, m, type) + 0.1f;
-				auto profit = (output - input - wage * type.get_base_workforce()) * (1.0f - rich_effect);
-				auto roi = profit / cost;
-
-				if((!lacking_input && !lacking_constr && (lacking_output || (profit / cost > 0.005f))) || profit / input > 1.00f)
-					desired_types.push_back(type.id);
-			} // END if building unlocked
-		}
-	}
-
-	// second pass: try to create factories which have a good profit margin
-	if(desired_types.empty()) {
-		for(auto type : state.world.in_factory_type) {
-			if(n.get_active_building(type) || type.get_is_available_from_start()) {
-				auto& inputs = type.get_inputs();
-				bool lacking_input = false;
-				bool lacking_output = m.get_demand_satisfaction(type.get_output()) < 0.98f;
-
-				for(uint32_t i = 0; i < economy::commodity_set::set_size; ++i) {
-					if(inputs.commodity_type[i]) {
-						if(m.get_demand_satisfaction(inputs.commodity_type[i]) < 0.5f)
-							lacking_input = true;
-					} else {
-						break;
-					}
-				}
-
-				auto& constr_cost = type.get_construction_costs();
-				auto lacking_constr = false;
-
-				for(uint32_t i = 0; i < economy::commodity_set::set_size; ++i) {
-					if(constr_cost.commodity_type[i]) {
-						if(m.get_demand_satisfaction(constr_cost.commodity_type[i]) < 0.1f)
-							lacking_constr = true;
-					} else {
-						break;
-					}
-				}
-
-				float cost = economy::factory_type_build_cost(state, n, pid, type, pop_project) + 0.1f;
-				float output = economy::factory_type_output_cost(state, n, m, type);
-				float input = economy::factory_type_input_cost(state, n, m, type) + 0.1f;
-				auto profit = (output - input - wage * type.get_base_workforce()) * (1.0f - rich_effect);
-				auto roi = profit / cost;
-
-				if(!lacking_input && !lacking_constr && profit / input > 0.3f && roi > 0.001f)
-					desired_types.push_back(type.id);
-			} // END if building unlocked
-		}
-	}
-}
-
-void get_state_craved_factory_types(sys::state& state, dcon::nation_id nid, dcon::market_id mid, dcon::province_id pid, std::vector<dcon::factory_type_id>& desired_types) {
-	assert(desired_types.empty());
-	auto n = dcon::fatten(state.world, nid);
-	auto m = dcon::fatten(state.world, mid);
-	auto sid = m.get_zone_from_local_market();
-	auto treasury = n.get_stockpiles(economy::money);
-	auto wage = state.world.province_get_labor_price(pid, economy::labor::basic_education) * 2.f;
-
-
-	if(desired_types.empty()) {
-		for(auto type : state.world.in_factory_type) {
-			if(n.get_active_building(type) || type.get_is_available_from_start()) {
-
-				float cost = economy::factory_type_build_cost(state, n, pid, type, false) + 0.1f;
-				float output = economy::factory_type_output_cost(state, n, m, type);
-				float input = economy::factory_type_input_cost(state, n, m, type) + 0.1f;
-
-				if((output - input - wage * type.get_base_workforce()) / input > 20.f)
-					desired_types.push_back(type.id);
-			} // END if building unlocked
-		}
-	}
-}
-
-void get_state_desired_factory_types(sys::state& state, dcon::nation_id nid, dcon::market_id mid, dcon::province_id pid, std::vector<dcon::factory_type_id>& desired_types) {
-	assert(desired_types.empty());
-	auto n = dcon::fatten(state.world, nid);
-	auto m = dcon::fatten(state.world, mid);
-	auto wage = state.world.province_get_labor_price(pid, economy::labor::basic_education) * 2.f;
-
-	auto treasury = n.get_stockpiles(economy::money);
-
-	// first pass: try to create factories which will pay back investment fast - in a year at most:
-	if(desired_types.empty()) {
-		for(auto type : state.world.in_factory_type) {
-			if(n.get_active_building(type) || type.get_is_available_from_start()) {
-				auto& inputs = type.get_inputs();
-				bool lacking_input = false;
-				bool lacking_output = m.get_demand_satisfaction(type.get_output()) < 0.98f;
-
-				for(uint32_t i = 0; i < economy::commodity_set::set_size; ++i) {
-					if(inputs.commodity_type[i]) {
-						if(m.get_demand_satisfaction(inputs.commodity_type[i]) < 0.5f)
-							lacking_input = true;
-					} else {
-						break;
-					}
-				}
-
-				float cost = economy::factory_type_build_cost(state, n, pid, type, false) + 0.1f;
-				float output = economy::factory_type_output_cost(state, n, m, type);
-				float input = economy::factory_type_input_cost(state, n, m, type) + 0.1f;
-
-				if((lacking_output || ((output - input - wage * type.get_base_workforce()) / cost < 365.f)))
-					desired_types.push_back(type.id);
-			} // END if building unlocked
-		}
-	}
-
-	// second pass: try to create factories which have a good profit margin
-	if(desired_types.empty()) {
-		for(auto type : state.world.in_factory_type) {
-			if(n.get_active_building(type) || type.get_is_available_from_start()) {
-				auto& inputs = type.get_inputs();
-				bool lacking_input = false;
-				bool lacking_output = m.get_demand_satisfaction(type.get_output()) < 0.98f;
-
-				for(uint32_t i = 0; i < economy::commodity_set::set_size; ++i) {
-					if(inputs.commodity_type[i]) {
-						if(m.get_demand_satisfaction(inputs.commodity_type[i]) < 0.5f)
-							lacking_input = true;
-					} else {
-						break;
-					}
-				}
-
-				float cost = economy::factory_type_build_cost(state, n, pid, type, false) + 0.1f;
-				float output = economy::factory_type_output_cost(state, n, m, type);
-				float input = economy::factory_type_input_cost(state, n, m, type) + 0.1f;
-				auto profitabilitymark = std::max(0.01f, cost * 10.f / treasury);
-
-				if((lacking_output || ((output - input - wage * type.get_base_workforce()) / input > profitabilitymark)))
-					desired_types.push_back(type.id);
-			} // END if building unlocked
-		}
-	}
-}
-
 int32_t future_rebels_in_nation(sys::state& state, dcon::nation_id n) {
 	auto total = 0;
 	for(auto fac : state.world.nation_get_rebellion_within(n)) {
@@ -1498,550 +236,6 @@ int16_t calculate_desired_army_size(sys::state& state, dcon::nation_id nation) {
 	return int16_t(std::clamp(ceil(total * double(factor)), ceil(0.1 * fid.get_recruitable_regiments()), 1.0 * fid.get_recruitable_regiments()));
 }
 
-void update_ai_econ_construction(sys::state& state) {
-	for(auto n : state.world.in_nation) {
-		// skip over: non ais, dead nations, and nations that aren't making money
-		if(n.get_owned_province_count() == 0 || !n.get_is_civilized())
-			continue;
-
-		if(n.get_is_player_controlled()) {
-			// to handle the logic of player building automation later
-			continue;
-		}
-
-		/*
-		if(n.get_spending_level() < 1.0f || n.get_last_treasury() >= n.get_stockpiles(economy::money))
-			continue;
-		*/
-
-		float treasury = n.get_stockpiles(economy::money);
-		float base_income = economy::estimate_daily_income(state, n);
-		float estimated_construction_costs = economy::estimate_construction_spending_from_budget(state, n, std::max(treasury, 1'000'000'000'000.f));
-
-		//if our army is too small, ignore buildings:
-		if(calculate_desired_army_size(state, n) * 0.4f > n.get_active_regiments())
-			continue;
-
-		float budget = treasury * 0.5f + base_income - estimated_construction_costs * 2.f;
-		float additional_expenses = 0.f;
-		float days_prepaid = 10.f;
-		auto rules = n.get_combined_issue_rules();
-
-		if(budget < 0.f) {
-			continue;
-		}
-
-		if((rules & issue_rule::expand_factory) != 0 || (rules & issue_rule::build_factory) != 0) {
-			// prepare a list of states
-			static std::vector<dcon::province_id> ordered_provinces;
-			ordered_provinces.clear();
-			for(auto p : n.get_province_ownership()) {
-				if(p.get_province().get_state_membership().get_capital().get_is_colonial() == false)
-					ordered_provinces.push_back(p.get_province());
-			}
-			std::sort(ordered_provinces.begin(), ordered_provinces.end(), [&](auto a, auto b) {
-				auto apop = state.world.province_get_demographics(a, demographics::total);
-				auto bpop = state.world.province_get_demographics(b, demographics::total);
-				if(apop != bpop)
-					return apop > bpop;
-				else
-					return a.index() < b.index();
-			});
-
-			// try to build
-			static::std::vector<dcon::factory_type_id> craved_types;
-
-			// desired types filled: try to construct or upgrade
-			if(!craved_types.empty()) {
-				if((rules & issue_rule::build_factory) == 0 && (rules & issue_rule::expand_factory) != 0) { // can't build -- by elimination, can upgrade
-
-				} else if((rules & issue_rule::build_factory) != 0) { // -- i.e. if building is possible
-					for(auto p : ordered_provinces) {
-						auto sid = state.world.province_get_state_membership(p);
-						auto market = state.world.state_instance_get_market_from_local_market(sid);
-
-						if(budget - additional_expenses <= 0.f)
-							break;
-
-						craved_types.clear();
-						get_state_craved_factory_types(state, n, market, p, craved_types);
-
-						// check -- either unemployed factory workers or no factory workers
-						auto pw_num = state.world.province_get_demographics(p,
-								demographics::to_key(state, state.culture_definitions.primary_factory_worker));
-						pw_num += state.world.province_get_demographics(p,
-								demographics::to_key(state, state.culture_definitions.secondary_factory_worker));
-						auto pw_employed = state.world.province_get_demographics(p,
-								demographics::to_employment_key(state, state.culture_definitions.primary_factory_worker));
-						pw_employed += state.world.province_get_demographics(p,
-								demographics::to_employment_key(state, state.culture_definitions.secondary_factory_worker));
-
-						if(pw_employed >= float(pw_num) * 2.5f && pw_num > 0.0f)
-							continue; // no spare workers
-
-						auto type_selection = craved_types[rng::get_random(state, uint32_t(n.id.index() + int32_t(budget))) % craved_types.size()];
-						assert(type_selection);
-
-						if(state.world.factory_type_get_is_coastal(type_selection) && !state.world.province_get_port_to(p))
-							continue;
-
-						bool already_in_progress = [&]() {
-							for(auto fc : state.world.province_get_factory_construction(p)) {
-								if(fc.get_type() == type_selection)
-									return true;
-							}
-							return false;
-							}();
-
-							if(already_in_progress)
-								continue;
-
-							// check: if present, try to upgrade
-							bool present_in_location = false;
-							bool under_cap = false;
-
-							for(auto fac : state.world.province_get_factory_location(p)) {
-								auto type = fac.get_factory().get_building_type();
-								if(type_selection == type) {
-									under_cap = economy::factory_total_employment_score(state, fac.get_factory()) < 0.9f;
-									present_in_location = true;
-									return;
-								}
-							}
-							if(under_cap) {
-								continue; // factory doesn't need to get larger
-							}
-
-							auto expected_item_cost = 0.f;
-							auto costs = state.world.factory_type_get_construction_costs(type_selection);
-							auto time = state.world.factory_type_get_construction_time(type_selection);
-							for(uint32_t i = 0; i < costs.set_size; ++i) {
-								if(costs.commodity_type[i]) {
-									expected_item_cost +=
-										costs.commodity_amounts[i]
-										* economy::price(state, market, costs.commodity_type[i])
-										/ float(time)
-										* days_prepaid;
-								} else {
-									break;
-								}
-							}
-
-							if(budget - additional_expenses - expected_item_cost <= 0.f)
-								continue;
-
-							if(present_in_location) {
-								if((rules & issue_rule::expand_factory) != 0 && economy::do_resource_potentials_allow_upgrade(state, n, p, type_selection)) {
-									auto new_up = fatten(state.world, state.world.force_create_factory_construction(p, n));
-									new_up.set_is_pop_project(false);
-									new_up.set_is_upgrade(true);
-									new_up.set_type(type_selection);
-
-									additional_expenses += expected_item_cost;
-								}
-								continue;
-							}
-
-							// else -- try to build -- must have room
-							auto num_factories = economy::province_factory_count(state, p);
-							if(num_factories < int32_t(state.defines.factories_per_state) && economy::do_resource_potentials_allow_construction(state, n, p, type_selection)) {
-								auto new_up = fatten(state.world, state.world.force_create_factory_construction(p, n));
-								new_up.set_is_pop_project(false);
-								new_up.set_is_upgrade(false);
-								new_up.set_type(type_selection);
-								additional_expenses += expected_item_cost;
-								continue;
-							} else {
-								// TODO: try to delete a factory here
-							}
-					} // END for(auto si : ordered_states) {
-				} // END if((rules & issue_rule::build_factory) == 0)
-			} // END if(!desired_types.empty()) {
-
-			// try to upgrade factories first:
-			if((rules & issue_rule::expand_factory) != 0) { // can't build -- by elimination, can upgrade
-				for(auto pid : ordered_provinces) {
-
-					auto sid = state.world.province_get_state_membership(pid);
-					auto market = state.world.state_instance_get_market_from_local_market(sid);
-
-					if(budget - additional_expenses <= 0.f)
-						break;
-
-					for(auto fac : state.world.province_get_factory_location(pid)) {
-						auto type = fac.get_factory().get_building_type();
-
-
-						auto unprofitable = fac.get_factory().get_unprofitable();
-
-						if(!unprofitable && economy::factory_total_employment_score(state, fac.get_factory()) >= 0.9f) {
-							// test if factory is already upgrading
-							auto ug_in_progress = false;
-							for(auto c : state.world.province_get_factory_construction(pid)) {
-								if(c.get_type() == type) {
-									ug_in_progress = true;
-									break;
-								}
-							}
-
-							auto expected_item_cost = 0.f;
-							auto& costs = state.world.factory_type_get_construction_costs(type);
-							auto& time = state.world.factory_type_get_construction_time(type);
-							for(uint32_t i = 0; i < costs.set_size; ++i) {
-								if(costs.commodity_type[i]) {
-									expected_item_cost +=
-										costs.commodity_amounts[i]
-										* economy::price(state, market, costs.commodity_type[i])
-										/ float(time)
-										* days_prepaid;
-								} else {
-									break;
-								}
-							}
-
-							if(budget - additional_expenses - expected_item_cost <= 0.f)
-								continue;
-
-							if(!ug_in_progress && economy::do_resource_potentials_allow_upgrade(state, n, pid, type)) {
-								auto new_up = fatten(state.world, state.world.force_create_factory_construction(pid, n));
-								new_up.set_is_pop_project(false);
-								new_up.set_is_upgrade(true);
-								new_up.set_type(type);
-
-								additional_expenses += expected_item_cost;
-							}
-						}
-					}
-				}
-			}
-
-			// try to build
-			static::std::vector<dcon::factory_type_id> desired_types;
-
-			// desired types filled: try to construct or upgrade
-			if(!desired_types.empty()) {				
-				if((rules & issue_rule::build_factory) == 0 && (rules & issue_rule::expand_factory) != 0) { // can't build -- by elimination, can upgrade
-					
-				} else if((rules & issue_rule::build_factory) != 0) { // -- i.e. if building is possible
-					for(auto pid : ordered_provinces) {
-
-						auto sid = state.world.province_get_state_membership(pid);
-						auto market = state.world.state_instance_get_market_from_local_market(sid);
-
-						if(budget - additional_expenses <= 0.f)
-							break;
-
-						desired_types.clear();
-						get_state_desired_factory_types(state, n, market, pid, desired_types);
-
-						// check -- either unemployed factory workers or no factory workers
-						auto pw_num = state.world.province_get_demographics(pid,
-								demographics::to_key(state, state.culture_definitions.primary_factory_worker));
-						pw_num += state.world.province_get_demographics(pid,
-								demographics::to_key(state, state.culture_definitions.secondary_factory_worker));
-						auto pw_employed = state.world.province_get_demographics(pid,
-								demographics::to_employment_key(state, state.culture_definitions.primary_factory_worker));
-						pw_employed += state.world.province_get_demographics(pid,
-								demographics::to_employment_key(state, state.culture_definitions.secondary_factory_worker));
-
-						if(pw_employed >= float(pw_num) * 2.5f && pw_num > 0.0f)
-							continue; // no spare workers
-
-						auto type_selection = desired_types[rng::get_random(state, uint32_t(n.id.index() + int32_t(budget))) % desired_types.size()];
-						assert(type_selection);
-
-						if(state.world.factory_type_get_is_coastal(type_selection) && !state.world.province_get_port_to(pid))
-							continue;
-
-						bool already_in_progress = [&]() {
-							for(auto p : state.world.province_get_factory_construction(pid)) {
-								if(p.get_type() == type_selection)
-									return true;
-							}
-							return false;
-						}();
-
-						if(already_in_progress)
-							continue;
-
-						// check: if present, try to upgrade
-						bool present_in_location = false;
-						bool under_cap = false;
-
-						for(auto fac : state.world.province_get_factory_location(pid)) {
-							auto type = fac.get_factory().get_building_type();
-							if(type_selection == type) {
-								under_cap = economy::factory_total_employment_score(state, fac.get_factory()) < 0.9f;
-								present_in_location = true;
-								return;
-							}
-						}
-						if(under_cap) {
-							continue; // factory doesn't need to get larger
-						}
-
-						auto expected_item_cost = 0.f;
-						auto costs = state.world.factory_type_get_construction_costs(type_selection);
-						auto time = state.world.factory_type_get_construction_time(type_selection);
-						for(uint32_t i = 0; i < costs.set_size; ++i) {
-							if(costs.commodity_type[i]) {
-								expected_item_cost +=
-									costs.commodity_amounts[i]
-									* economy::price(state, market, costs.commodity_type[i])
-									/ float(time)
-									* days_prepaid;
-							} else {
-								break;
-							}
-						}
-
-						if(budget - additional_expenses - expected_item_cost <= 0.f)
-							continue;
-
-						if(present_in_location) {
-							if((rules & issue_rule::expand_factory) != 0 && economy::do_resource_potentials_allow_upgrade(state, n, pid, type_selection)) {
-								auto new_up = fatten(state.world, state.world.force_create_factory_construction(pid, n));
-								new_up.set_is_pop_project(false);
-								new_up.set_is_upgrade(true);
-								new_up.set_type(type_selection);
-
-								additional_expenses += expected_item_cost;
-							}
-							continue;
-						}
-
-						// else -- try to build -- must have room
-						int32_t num_factories = economy::province_factory_count(state, pid);
-						if(num_factories < int32_t(state.defines.factories_per_state) && economy::do_resource_potentials_allow_construction(state, n, pid, type_selection)) {
-							auto new_up = fatten(state.world, state.world.force_create_factory_construction(pid, n));
-							new_up.set_is_pop_project(false);
-							new_up.set_is_upgrade(false);
-							new_up.set_type(type_selection);
-							additional_expenses += expected_item_cost;
-							continue;
-						} else {
-							// TODO: try to delete a factory here
-						}
-					} // END for(auto si : ordered_states) {
-				} // END if((rules & issue_rule::build_factory) == 0)
-			} // END if(!desired_types.empty()) {
-		} // END  if((rules & issue_rule::expand_factory) != 0 || (rules & issue_rule::build_factory) != 0)
-
-		// AI has smarter desired army size, this check stops economic devt.
-		//if(0.9f * n.get_recruitable_regiments() > n.get_active_regiments())
-		//	continue;
-
-		static std::vector<dcon::province_id> project_provs;
-		project_provs.clear();
-
-		// try naval bases
-		if(budget - additional_expenses >= 0.f) {
-			project_provs.clear();
-			for(auto o : n.get_province_ownership()) {
-				if(!o.get_province().get_is_coast())
-					continue;
-				if(n != o.get_province().get_nation_from_province_control())
-					continue;				
-
-				if(military::province_is_under_siege(state, o.get_province()))
-					continue;
-				if(o.get_province().get_building_level(uint8_t(economy::province_building_type::naval_base)) == 0 && o.get_province().get_state_membership().get_naval_base_is_taken())
-					continue;
-
-				int32_t current_lvl = o.get_province().get_building_level(uint8_t(economy::province_building_type::naval_base));
-				int32_t max_local_lvl = n.get_max_building_level(uint8_t(economy::province_building_type::naval_base));
-				int32_t min_build = int32_t(o.get_province().get_modifier_values(sys::provincial_mod_offsets::min_build_naval_base));
-
-				if(max_local_lvl - current_lvl - min_build <= 0)
-					continue;
-
-				if(!province::has_naval_base_being_built(state, o.get_province())) {
-					project_provs.push_back(o.get_province().id);
-				}
-			}
-
-			auto cap = n.get_capital();
-			std::sort(project_provs.begin(), project_provs.end(), [&](dcon::province_id a, dcon::province_id b) {
-				auto a_dist = province::sorting_distance(state, a, cap);
-				auto b_dist = province::sorting_distance(state, b, cap);
-				if(a_dist != b_dist)
-					return a_dist < b_dist;
-				else
-					return a.index() < b.index();
-			});
-			if(!project_provs.empty()) {
-				auto si = state.world.province_get_state_membership(project_provs[0]);
-				auto market = state.world.state_instance_get_market_from_local_market(si);
-
-				// avoid overbuilding!
-
-				auto expected_item_cost = 0.f;
-				auto& costs = state.economy_definitions.building_definitions[int32_t(economy::province_building_type::naval_base)].cost;
-				auto& time = state.economy_definitions.building_definitions[int32_t(economy::province_building_type::naval_base)].time;
-				for(uint32_t i = 0; i < costs.set_size; ++i) {
-					if(costs.commodity_type[i]) {
-						expected_item_cost +=
-							costs.commodity_amounts[i]
-							* economy::price(state, market, costs.commodity_type[i])
-							/ float(time)
-							* days_prepaid;
-					} else {
-						break;
-					}
-				}
-
-				if(budget - additional_expenses - expected_item_cost <= 0.f)
-					continue;
-
-				if(si)
-					si.set_naval_base_is_taken(true);
-				auto new_rr = fatten(state.world, state.world.force_create_province_building_construction(project_provs[0], n));
-				new_rr.set_is_pop_project(false);
-				new_rr.set_type(uint8_t(economy::province_building_type::naval_base));
-				additional_expenses += expected_item_cost;
-			}
-		}
-
-		// try railroads
-		const struct {
-			bool buildable;
-			economy::province_building_type type;
-			dcon::provincial_modifier_value mod;
-		} econ_buildable[3] = {
-			{ (rules & issue_rule::build_railway) != 0, economy::province_building_type::railroad, sys::provincial_mod_offsets::min_build_railroad },
-			{ (rules & issue_rule::build_bank) != 0 && state.economy_definitions.building_definitions[uint32_t(economy::province_building_type::bank)].defined, economy::province_building_type::bank, sys::provincial_mod_offsets::min_build_bank },
-			{ (rules & issue_rule::build_university) != 0 && state.economy_definitions.building_definitions[uint32_t(economy::province_building_type::university)].defined, economy::province_building_type::university, sys::provincial_mod_offsets::min_build_university }
-		};
-		for(auto i = 0; i < 3; i++) {
-			if(econ_buildable[i].buildable && budget - additional_expenses > 0) {
-				project_provs.clear();
-				for(auto o : n.get_province_ownership()) {
-					if(n != o.get_province().get_nation_from_province_control())
-						continue;
-					if(military::province_is_under_siege(state, o.get_province()))
-						continue;
-					int32_t current_lvl = state.world.province_get_building_level(o.get_province(), uint8_t(econ_buildable[i].type));
-					int32_t max_local_lvl = state.world.nation_get_max_building_level(n, uint8_t(econ_buildable[i].type));
-					int32_t min_build = int32_t(state.world.province_get_modifier_values(o.get_province(), econ_buildable[i].mod));
-					if(max_local_lvl - current_lvl - min_build <= 0)
-						continue;
-					if(!province::has_province_building_being_built(state, o.get_province(), econ_buildable[i].type)) {
-						project_provs.push_back(o.get_province().id);
-					}
-				}
-				auto cap = n.get_capital();
-				std::sort(project_provs.begin(), project_provs.end(), [&](dcon::province_id a, dcon::province_id b) {
-					auto a_dist = province::sorting_distance(state, a, cap);
-					auto b_dist = province::sorting_distance(state, b, cap);
-					if(a_dist != b_dist)
-						return a_dist < b_dist;
-					else
-						return a.index() < b.index();
-				});
-				for(uint32_t j = 0; j < project_provs.size() && budget - additional_expenses > 0; ++j) {
-					auto sid = state.world.province_get_state_membership(project_provs[j]);
-					auto market = state.world.state_instance_get_market_from_local_market(sid);
-
-					// avoid overbuilding!
-
-					auto expected_item_cost = 0.f;
-					auto& costs = state.economy_definitions.building_definitions[uint8_t(econ_buildable[i].type)].cost;
-					auto& time = state.economy_definitions.building_definitions[uint8_t(econ_buildable[i].type)].time;
-					for(uint32_t k = 0; k < costs.set_size; ++k) {
-						if(costs.commodity_type[k]) {
-							expected_item_cost +=
-								costs.commodity_amounts[k]
-								* economy::price(state, market, costs.commodity_type[k])
-								/ float(time)
-								* days_prepaid;
-						} else {
-							break;
-						}
-					}
-
-					if(budget - additional_expenses - expected_item_cost <= 0.f)
-						continue;
-
-					auto new_proj = fatten(state.world, state.world.force_create_province_building_construction(project_provs[j], n));
-					new_proj.set_is_pop_project(false);
-					new_proj.set_type(uint8_t(econ_buildable[i].type));
-					additional_expenses += expected_item_cost;
-				}
-			}
-		}
-
-		// AI has smarter desired army size, this check stops economic devt.
-		//if(0.95f * n.get_recruitable_regiments() > n.get_active_regiments())
-		//	continue;
-
-		// try forts
-		if(budget - additional_expenses > 0.f) {
-			project_provs.clear();
-
-			for(auto o : n.get_province_ownership()) {
-				if(n != o.get_province().get_nation_from_province_control())
-					continue;
-
-				if(military::province_is_under_siege(state, o.get_province()))
-					continue;
-
-				int32_t current_lvl = state.world.province_get_building_level(o.get_province(), uint8_t(economy::province_building_type::fort));
-				int32_t max_local_lvl = state.world.nation_get_max_building_level(n, uint8_t(economy::province_building_type::fort));
-				int32_t min_build = int32_t(state.world.province_get_modifier_values(o.get_province(), sys::provincial_mod_offsets::min_build_fort));
-
-				if(max_local_lvl - current_lvl - min_build <= 0)
-					continue;
-
-				if(!province::has_fort_being_built(state, o.get_province())) {
-					project_provs.push_back(o.get_province().id);
-				}
-			}
-
-			auto cap = n.get_capital();
-			std::sort(project_provs.begin(), project_provs.end(), [&](dcon::province_id a, dcon::province_id b) {
-				auto a_dist = province::sorting_distance(state, a, cap);
-				auto b_dist = province::sorting_distance(state, b, cap);
-				if(a_dist != b_dist)
-					return a_dist < b_dist;
-				else
-					return a.index() < b.index();
-			});
-
-			for(uint32_t i = 0; i < project_provs.size() && budget - additional_expenses > 0.f; ++i) {
-
-				auto sid = state.world.province_get_state_membership(project_provs[i]);
-				auto market = state.world.state_instance_get_market_from_local_market(sid.id);
-
-				// avoid overbuilding!
-
-				auto expected_item_cost = 0.f;
-				auto& costs = state.economy_definitions.building_definitions[uint8_t(economy::province_building_type::fort)].cost;
-				auto& time = state.economy_definitions.building_definitions[uint8_t(economy::province_building_type::fort)].time;
-				for(uint32_t k = 0; k < costs.set_size; ++k) {
-					if(costs.commodity_type[k]) {
-						expected_item_cost +=
-							costs.commodity_amounts[k]
-							* economy::price(state, market, costs.commodity_type[k])
-							* days_prepaid
-							* 100000.f;
-						// forts are very bad investment and demand volatile goods,
-						// so build them if AI is really rich and has no idea where how to spend money
-					} else {
-						break;
-					}
-				}
-
-				if(budget - additional_expenses - expected_item_cost <= 0.f)
-					continue;
-
-				auto new_rr = fatten(state.world, state.world.force_create_province_building_construction(project_provs[i], n));
-				new_rr.set_is_pop_project(false);
-				new_rr.set_type(uint8_t(economy::province_building_type::fort));
-				additional_expenses += expected_item_cost;
-			}
-		}
-	}
-}
 
 void update_ai_colonial_investment(sys::state& state) {
 	static std::vector<dcon::state_definition_id> investments;
@@ -2241,1991 +435,6 @@ void take_reforms(sys::state& state) {
 		}
 	}
 }
-
-bool will_be_crisis_primary_attacker(sys::state& state, dcon::nation_id n) {
-
-	auto first_wg = state.crisis_attacker_wargoals.at(0);
-	if(first_wg.cb == state.military_definitions.crisis_colony) {
-		auto colonizers = state.world.state_definition_get_colonization(first_wg.state);
-		if(colonizers.end() - colonizers.begin() < 2)
-			return false;
-
-		auto defending_colonizer = (*(colonizers.begin() + 1)).get_colonizer();
-		if(state.world.nation_get_in_sphere_of(defending_colonizer) == n)
-			return false;
-		if(state.world.nation_get_ai_rival(n) == defending_colonizer
-			|| (defending_colonizer.get_in_sphere_of() && !nations::are_allied(state, n, defending_colonizer) && state.world.nation_get_ai_rival(n) == defending_colonizer.get_in_sphere_of())) {
-			return true;
-		} else {
-			if(state.primary_crisis_defender && state.world.nation_get_ai_rival(n) == state.primary_crisis_defender)
-				return true;
-			return false;
-		}
-	} else if(first_wg.cb == state.military_definitions.liberate) {
-		auto state_owner = first_wg.target_nation;
-		auto liberated = state.world.national_identity_get_nation_from_identity_holder(first_wg.wg_tag);
-
-		if(state_owner == n) //don't shoot ourselves
-			return false;
-		if(liberated == n) //except when we are shooting someone else
-			return true;
-		if(state.world.nation_get_in_sphere_of(state_owner) == n || nations::are_allied(state, n, state_owner))
-			return false;
-		if(state.world.nation_get_ai_rival(n) == state_owner)
-			return true;
-		if(state.world.nation_get_in_sphere_of(state_owner) && state.world.nation_get_ai_rival(n) == state.world.nation_get_in_sphere_of(state_owner))
-			return true;
-		if(state.world.nation_get_in_sphere_of(liberated) == n || nations::are_allied(state, n, liberated))
-			return true;
-		if(state.primary_crisis_defender && state.world.nation_get_ai_rival(n) == state.primary_crisis_defender)
-			return true;
-		return false;
-	} else {
-		return false;
-	}
-}
-bool will_be_crisis_primary_defender(sys::state& state, dcon::nation_id n) {
-	auto first_wg = state.crisis_attacker_wargoals.at(0);
-	if(first_wg.cb == state.military_definitions.crisis_colony) {
-		auto colonizers = state.world.state_definition_get_colonization(first_wg.state);
-		if(colonizers.end() - colonizers.begin() < 2)
-			return false;
-
-		auto attacking_colonizer = (*colonizers.begin()).get_colonizer();
-
-		if(state.world.nation_get_in_sphere_of(attacking_colonizer) == n)
-			return false;
-		if(state.world.nation_get_ai_rival(n) == attacking_colonizer
-			|| (attacking_colonizer.get_in_sphere_of() && !nations::are_allied(state, n, attacking_colonizer) && state.world.nation_get_ai_rival(n) == attacking_colonizer.get_in_sphere_of())
-			|| state.world.nation_get_ai_rival(n) == state.primary_crisis_attacker) {
-			return true;
-		} else {
-			if(state.primary_crisis_attacker && state.world.nation_get_ai_rival(n) == state.primary_crisis_attacker)
-				return true;
-			return false;
-		}
-	} else if(first_wg.cb == state.military_definitions.liberate) {
-		auto state_owner = first_wg.target_nation;
-		auto liberated = state.world.national_identity_get_nation_from_identity_holder(first_wg.wg_tag);
-
-		if(state_owner == n) //don't shoot ourselves
-			return false;
-		if(liberated == n) //except when we are shooting someone else
-			return true;
-		if(state.world.nation_get_in_sphere_of(liberated) == n || nations::are_allied(state, n, liberated))
-			return false;
-		if(state.world.nation_get_ai_rival(n) == liberated)
-			return true;
-		if(state.world.nation_get_in_sphere_of(liberated) && state.world.nation_get_ai_rival(n) == state.world.nation_get_in_sphere_of(liberated))
-			return true;
-		if(state.world.nation_get_in_sphere_of(state_owner) == n || nations::are_allied(state, n, state_owner))
-			return true;
-		if(state.primary_crisis_attacker && state.world.nation_get_ai_rival(n) == state.primary_crisis_attacker)
-			return true;
-		return false;
-	} else {
-		return false;
-	}
-}
-
-struct crisis_str {
-	float attacker = 0.0f;
-	float defender = 0.0f;
-	bool defender_win = false;
-	bool attacker_win = false;
-	bool fast_victory = false;
-};
-
-crisis_str estimate_crisis_str(sys::state& state) {
-	float atotal = 0.0f;
-	float dtotal = 0.0f;
-
-	if(state.crisis_attacker && state.crisis_attacker != state.primary_crisis_attacker) {
-		atotal += estimate_strength(state, state.crisis_attacker);
-	}
-	if(state.crisis_defender && state.crisis_defender != state.primary_crisis_defender) {
-		dtotal += estimate_strength(state, state.crisis_defender);
-	}
-	for(auto& i : state.crisis_participants) {
-		if(!i.id)
-			break;
-		if(!i.merely_interested) {
-			if(i.supports_attacker) {
-				atotal += estimate_strength(state, i.id);
-			} else {
-				dtotal += estimate_strength(state, i.id);
-			}
-		}
-	}
-
-	auto necessary_atk_win_ratio = state.defines.alice_crisis_necessary_base_win_ratio;
-	auto necessary_def_win_ratio = state.defines.alice_crisis_necessary_base_win_ratio;
-	auto necessary_fast_win_ratio = state.defines.alice_crisis_necessary_base_fast_win_ratio;
-
-	if(!state.world.nation_get_is_civilized(state.crisis_defender)) {
-		necessary_atk_win_ratio += state.defines.alice_crisis_unciv_stubbornness;
-		necessary_fast_win_ratio += state.defines.alice_crisis_unciv_stubbornness;
-	}
-	for(auto wg : state.crisis_attacker_wargoals) {
-		if(!wg.cb) {
-			break;
-		}
-
-		necessary_atk_win_ratio += state.defines.alice_crisis_per_wg_ratio;
-		necessary_fast_win_ratio += state.defines.alice_crisis_per_wg_ratio;
-	}
-	for(auto wg : state.crisis_defender_wargoals) {
-		if(!wg.cb) {
-			break;
-		}
-
-		necessary_def_win_ratio += state.defines.alice_crisis_per_wg_ratio;
-		necessary_fast_win_ratio += state.defines.alice_crisis_per_wg_ratio;
-	}
-
-	auto def_victory = dtotal > atotal * necessary_def_win_ratio;
-	auto atk_victory = atotal > dtotal * necessary_atk_win_ratio;
-	auto fast_victory = (dtotal / atotal > necessary_fast_win_ratio) || (atotal / dtotal > necessary_fast_win_ratio);
-	if(fast_victory) {
-		atk_victory = atotal > dtotal;
-		def_victory = dtotal > atotal;
-	}
-
-	return crisis_str{ atotal, dtotal, def_victory, atk_victory, fast_victory};
-}
-
-bool will_join_crisis_with_offer(sys::state& state, dcon::nation_id n, sys::full_wg offer) {
-	if(offer.target_nation == state.world.nation_get_ai_rival(n))
-		return true;
-	auto offer_bits = state.world.cb_type_get_type_bits(offer.cb);
-	if((offer_bits & (military::cb_flag::po_demand_state | military::cb_flag::po_annex)) != 0)
-		return true;
-
-	// GPs accept unequal treaties as crisis offer
-	if(state.world.nation_get_is_great_power(n) && (offer_bits & (military::cb_flag::po_unequal_treaty)) != 0)
-		return true;
-	return false;
-}
-
-bool ai_offer_cb(sys::state& state, dcon::cb_type_id t) {
-	auto offer_bits = state.world.cb_type_get_type_bits(t);
-	if((offer_bits & (military::cb_flag::po_demand_state | military::cb_flag::po_annex)) != 0)
-		return false;
-	if((offer_bits & military::cb_flag::all_allowed_states) != 0)
-		return false;
-	if(military::cb_requires_selection_of_a_liberatable_tag(state, t))
-		return false;
-	if(military::cb_requires_selection_of_a_valid_nation(state, t))
-		return false;
-	return true;
-}
-
-/* Prioritize states inside the nation */
-void state_target_list(std::vector<dcon::state_instance_id>& result, sys::state& state, dcon::nation_id for_nation, dcon::nation_id within) {
-	result.clear();
-	for(auto si : state.world.nation_get_state_ownership(within)) {
-		result.push_back(si.get_state().id);
-	}
-
-	auto distance_from = state.world.nation_get_capital(for_nation).id;
-	int32_t first = 0;
-
-	if(state.world.get_nation_adjacency_by_nation_adjacency_pair(for_nation, within)) {
-		int32_t last = int32_t(result.size());
-		while(first < last - 1) {
-			while(first < last && province::state_borders_nation(state, for_nation, result[first])) {
-				++first;
-			}
-			while(first < last - 1 && !province::state_borders_nation(state, for_nation, result[last - 1])) {
-				--last;
-			}
-			if(first < last - 1) {
-				std::swap(result[first], result[last - 1]);
-				++first;
-				--last;
-			}
-		}
-
-		std::sort(result.begin(), result.begin() + first, [&](dcon::state_instance_id a, dcon::state_instance_id b) {
-			auto a_distance = province::sorting_distance(state, state.world.state_instance_get_capital(a), distance_from);
-			auto b_distance = province::sorting_distance(state, state.world.state_instance_get_capital(b), distance_from);
-			if(a_distance != b_distance)
-				return a_distance < b_distance;
-			else
-				return a.index() < b.index();
-		});
-	}
-	if(state.world.nation_get_total_ports(for_nation) > 0 && state.world.nation_get_total_ports(within) > 0) {
-		int32_t last = int32_t(result.size());
-		while(first < last - 1) {
-			while(first < last && province::state_is_coastal(state, result[first])) {
-				++first;
-			}
-			while(first < last - 1 && !province::state_is_coastal(state, result[last - 1])) {
-				--last;
-			}
-			if(first < last - 1) {
-				std::swap(result[first], result[last - 1]);
-				++first;
-				--last;
-			}
-		}
-		std::sort(result.begin(), result.begin() + first, [&](dcon::state_instance_id a, dcon::state_instance_id b) {
-			auto a_distance = province::sorting_distance(state, state.world.state_instance_get_capital(a), distance_from);
-			auto b_distance = province::sorting_distance(state, state.world.state_instance_get_capital(b), distance_from);
-			if(a_distance != b_distance)
-				return a_distance < b_distance;
-			else
-				return a.index() < b.index();
-		});
-	}
-	if(first < int32_t(result.size())) {
-		std::sort(result.begin() + first, result.end(), [&](dcon::state_instance_id a, dcon::state_instance_id b) {
-			auto a_distance = province::sorting_distance(state, state.world.state_instance_get_capital(a), distance_from);
-			auto b_distance = province::sorting_distance(state, state.world.state_instance_get_capital(b), distance_from);
-			if(a_distance != b_distance)
-				return a_distance < b_distance;
-			else
-				return a.index() < b.index();
-		});
-	}
-}
-
-void update_crisis_leaders(sys::state& state) {
-	if(state.current_crisis_state == sys::crisis_state::inactive) {
-		return;
-	}
-
-	auto str_est = estimate_crisis_str(state);
-
-	if(state.crisis_temperature > 75.f || str_est.fast_victory) { // make peace offer
-		auto any_victory = str_est.attacker_win || str_est.defender_win;
-		auto defender_victory = str_est.defender_win;
-
-		auto will_propose_peace = state.world.nation_get_is_player_controlled(state.primary_crisis_attacker) == false && any_victory;
-
-		auto primary_wg = state.crisis_attacker_wargoals[0];
-
-		// Defender should never agree to full annex
-		if(state.crisis_defender == state.primary_crisis_defender) {
-			auto bits = state.world.cb_type_get_type_bits(primary_wg.cb);
-			if((bits & military::cb_flag::po_annex) != 0)
-				will_propose_peace = false;
-		}
-
-		if(state.crisis_temperature <= 20.f) {
-			will_propose_peace = false;
-		}
-		
-		if(will_propose_peace) {
-			assert(command::can_start_crisis_peace_offer(state, state.primary_crisis_attacker, defender_victory));
-			command::execute_start_crisis_peace_offer(state, state.primary_crisis_attacker, defender_victory);
-			auto pending = state.world.nation_get_peace_offer_from_pending_peace_offer(state.primary_crisis_attacker);
-
-			auto wargoalslist = (defender_victory) ? state.crisis_defender_wargoals : state.crisis_attacker_wargoals;
-			for(auto swg : wargoalslist) {
-				if(!swg.cb) {
-					break;
-				}
-				auto wg = fatten(state.world, state.world.create_wargoal());
-				wg.set_peace_offer_from_peace_offer_item(pending);
-				wg.set_added_by(swg.added_by);
-				wg.set_associated_state(swg.state);
-				wg.set_associated_tag(swg.wg_tag);
-				wg.set_secondary_nation(swg.secondary_nation);
-				wg.set_target_nation(swg.target_nation);
-				wg.set_type(swg.cb);
-				assert(command::can_add_to_crisis_peace_offer(state, state.primary_crisis_attacker, swg.added_by, swg.target_nation,
-					swg.cb, swg.state, swg.wg_tag, swg.secondary_nation));
-			}
-			assert(command::can_send_crisis_peace_offer(state, state.primary_crisis_attacker));
-			command::execute_send_crisis_peace_offer(state, state.primary_crisis_attacker);
-		} else if(state.world.nation_get_is_player_controlled(state.primary_crisis_defender) == false && any_victory) {
-			assert(command::can_start_crisis_peace_offer(state, state.primary_crisis_defender, !defender_victory));
-			command::execute_start_crisis_peace_offer(state, state.primary_crisis_defender, !defender_victory);
-			auto pending = state.world.nation_get_peace_offer_from_pending_peace_offer(state.primary_crisis_defender);
-
-			auto wargoalslist = (defender_victory) ? state.crisis_defender_wargoals : state.crisis_attacker_wargoals;
-			for(auto swg : wargoalslist) {
-				if(!swg.cb) {
-					break;
-				}
-				assert(command::can_add_to_crisis_peace_offer(state, state.primary_crisis_defender, swg.added_by, swg.target_nation,
-					swg.cb, swg.state, swg.wg_tag, swg.secondary_nation));
-				auto wg = fatten(state.world, state.world.create_wargoal());
-				wg.set_peace_offer_from_peace_offer_item(pending);
-				wg.set_added_by(swg.added_by);
-				wg.set_associated_state(swg.state);
-				wg.set_associated_tag(swg.wg_tag);
-				wg.set_secondary_nation(swg.secondary_nation);
-				wg.set_target_nation(swg.target_nation);
-				wg.set_type(swg.cb);
-				
-			}
-			assert(command::can_send_crisis_peace_offer(state, state.primary_crisis_defender));
-			command::execute_send_crisis_peace_offer(state, state.primary_crisis_defender);
-		}
-	} else if(state.crisis_temperature > 20.f) { // recruit nations
-		if(str_est.attacker < str_est.defender && state.world.nation_get_is_player_controlled(state.primary_crisis_attacker) == false) {
-			for(auto& par : state.crisis_participants) {
-				if(!par.id)
-					break;
-				if(par.merely_interested) {
-					auto other_cbs = state.world.nation_get_available_cbs(par.id);
-					dcon::cb_type_id offer_cb;
-					dcon::nation_id target;
-
-					[&]() {
-						for(auto& op_par : state.crisis_participants) {
-							if(!op_par.id)
-								break;
-							if(!op_par.merely_interested && op_par.supports_attacker == false) {
-								for(auto& cb : other_cbs) {
-									if(cb.target == op_par.id && ai_offer_cb(state, cb.cb_type) && military::cb_conditions_satisfied(state, par.id, op_par.id, cb.cb_type)) {
-										offer_cb = cb.cb_type;
-										target = op_par.id;
-										return;
-									}
-								}
-								for(auto cb : state.world.in_cb_type) {
-									if((cb.get_type_bits() & military::cb_flag::always) != 0) {
-										if(ai_offer_cb(state, cb) && military::cb_conditions_satisfied(state, par.id, op_par.id, cb)) {
-											offer_cb = cb;
-											target = op_par.id;
-											return;
-										}
-									}
-								}
-							}
-						}
-						}();
-
-						if(offer_cb) {
-							if(military::cb_requires_selection_of_a_state(state, offer_cb)) {
-								std::vector < dcon::state_instance_id> potential_states;
-								state_target_list(potential_states, state, par.id, target);
-								for(auto s : potential_states) {
-									if(military::cb_instance_conditions_satisfied(state, par.id, target, offer_cb, state.world.state_instance_get_definition(s), dcon::national_identity_id{}, dcon::nation_id{})) {
-
-										diplomatic_message::message m;
-										memset(&m, 0, sizeof(diplomatic_message::message));
-										m.to = par.id;
-										m.from = state.primary_crisis_attacker;
-										m.data.crisis_offer.added_by = m.to;
-										m.data.crisis_offer.target_nation = target;
-										m.data.crisis_offer.secondary_nation = dcon::nation_id{};
-										m.data.crisis_offer.state = state.world.state_instance_get_definition(s);
-										m.data.crisis_offer.wg_tag = dcon::national_identity_id{};
-										m.data.crisis_offer.cb = offer_cb;
-										m.type = diplomatic_message::type::take_crisis_side_offer;
-										diplomatic_message::post(state, m);
-
-										break;
-									}
-								}
-							} else {
-								diplomatic_message::message m;
-								memset(&m, 0, sizeof(diplomatic_message::message));
-								m.to = par.id;
-								m.from = state.primary_crisis_attacker;
-								m.data.crisis_offer.added_by = m.to;
-								m.data.crisis_offer.target_nation = target;
-								m.data.crisis_offer.secondary_nation = dcon::nation_id{};
-								m.data.crisis_offer.state = dcon::state_definition_id{};
-								m.data.crisis_offer.wg_tag = dcon::national_identity_id{};
-								m.data.crisis_offer.cb = offer_cb;
-								m.type = diplomatic_message::type::take_crisis_side_offer;
-								diplomatic_message::post(state, m);
-							}
-						}
-				}
-			}
-		} else if(str_est.attacker < str_est.defender && state.world.nation_get_is_player_controlled(state.primary_crisis_defender) == false) {
-			for(auto& par : state.crisis_participants) {
-				if(!par.id)
-					break;
-				if(par.merely_interested) {
-					auto other_cbs = state.world.nation_get_available_cbs(par.id);
-					dcon::cb_type_id offer_cb;
-					dcon::nation_id target;
-
-					[&]() {
-						for(auto& op_par : state.crisis_participants) {
-							if(!op_par.id)
-								break;
-							if(!op_par.merely_interested && op_par.supports_attacker == true) {
-								for(auto& cb : other_cbs) {
-									if(cb.target == op_par.id && ai_offer_cb(state, cb.cb_type) && military::cb_conditions_satisfied(state, par.id, op_par.id, cb.cb_type)) {
-										offer_cb = cb.cb_type;
-										target = op_par.id;
-										return;
-									}
-								}
-								for(auto cb : state.world.in_cb_type) {
-									if((cb.get_type_bits() & military::cb_flag::always) != 0) {
-										if(ai_offer_cb(state, cb) && military::cb_conditions_satisfied(state, par.id, op_par.id, cb)) {
-											offer_cb = cb;
-											target = op_par.id;
-											return;
-										}
-									}
-								}
-							}
-						}
-						}();
-
-						if(offer_cb) {
-							if(military::cb_requires_selection_of_a_state(state, offer_cb)) {
-								std::vector < dcon::state_instance_id> potential_states;
-								state_target_list(potential_states, state, par.id, target);
-								for(auto s : potential_states) {
-									if(military::cb_instance_conditions_satisfied(state, par.id, target, offer_cb, state.world.state_instance_get_definition(s), dcon::national_identity_id{}, dcon::nation_id{})) {
-
-										diplomatic_message::message m;
-										memset(&m, 0, sizeof(diplomatic_message::message));
-										m.to = par.id;
-										m.from = state.primary_crisis_defender;
-										m.data.crisis_offer.added_by = m.to;
-										m.data.crisis_offer.target_nation = target;
-										m.data.crisis_offer.secondary_nation = dcon::nation_id{};
-										m.data.crisis_offer.state = state.world.state_instance_get_definition(s);
-										m.data.crisis_offer.wg_tag = dcon::national_identity_id{};
-										m.data.crisis_offer.cb = offer_cb;
-										m.type = diplomatic_message::type::take_crisis_side_offer;
-										diplomatic_message::post(state, m);
-
-										break;
-									}
-								}
-							} else {
-								diplomatic_message::message m;
-								memset(&m, 0, sizeof(diplomatic_message::message));
-								m.to = par.id;
-								m.from = state.primary_crisis_defender;
-								m.data.crisis_offer.added_by = m.to;
-								m.data.crisis_offer.target_nation = target;
-								m.data.crisis_offer.secondary_nation = dcon::nation_id{};
-								m.data.crisis_offer.state = dcon::state_definition_id{};
-								m.data.crisis_offer.wg_tag = dcon::national_identity_id{};
-								m.data.crisis_offer.cb = offer_cb;
-								m.type = diplomatic_message::type::take_crisis_side_offer;
-								diplomatic_message::post(state, m);
-							}
-						}
-				}
-			}
-		}
-	}
-}
-
-bool will_accept_crisis_peace_offer(sys::state& state, dcon::nation_id to, bool is_concession, bool missing_wg) {
-
-	auto primary_wg = state.crisis_attacker_wargoals[0];
-
-	// Defender should never agree to full annex
-	if(state.crisis_defender == state.primary_crisis_defender && to == state.crisis_defender) {
-		auto bits = state.world.cb_type_get_type_bits(primary_wg.cb);
-			if((bits & military::cb_flag::po_annex) != 0)
-				return false;
-	}
-
-	if(state.crisis_temperature <= 20.f) {
-		return false;
-	}
-
-	auto str_est = estimate_crisis_str(state);
-
-	if(state.crisis_temperature > 75.f || str_est.fast_victory) {
-		if(to == state.primary_crisis_attacker) {
-			if(is_concession) {
-				return true;
-			}
-			if(str_est.defender_win) {
-				return true;
-			}
-
-			return false;
-		} else if(to == state.primary_crisis_defender) {
-			if(is_concession)
-				return true;
-
-			if(str_est.attacker_win)
-				return true;
-
-			return false;
-		}
-	}
-	return false;
-}
-
-bool will_accept_crisis_peace_offer(sys::state& state, dcon::nation_id to, dcon::peace_offer_id peace) {
-	auto primary_wg = state.crisis_attacker_wargoals[0];
-
-	// Defender should never agree to full annex
-	if(state.crisis_defender == state.primary_crisis_defender && to == state.crisis_defender) {
-		auto bits = state.world.cb_type_get_type_bits(primary_wg.cb);
-		if((bits & military::cb_flag::po_annex) != 0)
-			return false;
-	}
-
-	if(state.crisis_temperature <= 20.f) {
-		return false;
-	}
-
-	auto str_est = estimate_crisis_str(state);
-
-	if(state.crisis_temperature > 75.f || str_est.fast_victory) {
-		if(to == state.primary_crisis_attacker) {
-			if(str_est.defender_win)
-				return true;
-
-			if(!state.world.peace_offer_get_is_concession(peace))
-				return false;
-
-			bool missing_wg = false;
-			for(auto swg : state.crisis_attacker_wargoals) {
-				bool found_wg = false;
-				for(auto item : state.world.peace_offer_get_peace_offer_item(peace)) {
-					auto wg = item.get_wargoal();
-					if(wg.get_type() == swg.cb && wg.get_associated_state() == swg.state && wg.get_added_by() == swg.added_by && wg.get_target_nation() == swg.target_nation &&
-						wg.get_associated_tag() == swg.wg_tag) {
-						found_wg = true; continue;
-					}
-				}
-
-				if(!found_wg) {
-					return false;
-				}
-			}
-			return true;
-
-		} else if(to == state.primary_crisis_defender) {
-			if(str_est.attacker_win)
-				return true;
-
-			if(!state.world.peace_offer_get_is_concession(peace))
-				return false;
-
-			bool missing_wg = false;
-			for(auto swg : state.crisis_defender_wargoals) {
-				bool found_wg = false;
-				for(auto item : state.world.peace_offer_get_peace_offer_item(peace)) {
-					auto wg = item.get_wargoal();
-					if(wg.get_type() == swg.cb && wg.get_associated_state() == swg.state && wg.get_added_by() == swg.added_by && wg.get_target_nation() == swg.target_nation &&
-						wg.get_associated_tag() == swg.wg_tag) {
-						found_wg = true; continue;
-					}
-				}
-
-				if(!found_wg) {
-					return false;
-				}
-			}
-
-			return true;
-		}
-	}
-	return false;
-}
-
-void update_war_intervention(sys::state& state) {
-	for(auto& gp : state.great_nations) {
-		if(state.world.nation_get_is_player_controlled(gp.nation) == false && state.world.nation_get_is_at_war(gp.nation) == false){
-			bool as_attacker = false;
-			dcon::war_id intervention_target;
-			[&]() {
-				for(auto w : state.world.in_war) {
-					//GPs will try to intervene in wars to protect smaller nations in the same cultural union
-					if(command::can_intervene_in_war(state, gp.nation, w, false)) {
-						auto par = state.world.war_get_primary_defender(w);
-						if(state.world.nation_get_primary_culture(gp.nation).get_group_from_culture_group_membership() == state.world.nation_get_primary_culture(par).get_group_from_culture_group_membership()
-							&& !nations::is_great_power(state, par)
-						){
-							intervention_target = w;
-							return;
-						}
-					}
-					if(w.get_is_great()) {
-						if(command::can_intervene_in_war(state, gp.nation, w, false)) {
-							for(auto par : w.get_war_participant()) {
-								if(par.get_is_attacker() && military::can_use_cb_against(state, gp.nation, par.get_nation())) {
-									intervention_target = w;
-									return;
-								}
-							}
-						}
-						if(command::can_intervene_in_war(state, gp.nation, w, true)) {
-							for(auto par : w.get_war_participant()) {
-								if(!par.get_is_attacker() && military::can_use_cb_against(state, gp.nation, par.get_nation())) {
-									intervention_target = w;
-									as_attacker = true;
-									return;
-								}
-							}
-						}
-					} else if(military::get_role(state, w, state.world.nation_get_ai_rival(gp.nation)) == military::war_role::attacker) {
-						if(command::can_intervene_in_war(state, gp.nation, w, false)) {
-							intervention_target = w;
-							return;
-						}
-					}
-				}
-				}();
-				if(intervention_target) {
-					assert(command::can_intervene_in_war(state, gp.nation, intervention_target, as_attacker));
-					command::execute_intervene_in_war(state, gp.nation, intervention_target, as_attacker);
-				}
-		}
-	}
-}
-
-struct possible_cb {
-	dcon::nation_id target;
-	dcon::nation_id secondary_nation;
-	dcon::national_identity_id associated_tag;
-	dcon::state_definition_id state_def;
-	dcon::cb_type_id cb;
-};
-
-/* Filter out target_states in the target nation */
-void place_instance_in_result(sys::state& state, std::vector<possible_cb>& result, dcon::nation_id n, dcon::nation_id target, dcon::cb_type_id cb, std::vector<dcon::state_instance_id> const& target_states) {
-	auto can_use = state.world.cb_type_get_can_use(cb);
-	auto allowed_substates = state.world.cb_type_get_allowed_substate_regions(cb);
-
-	if(allowed_substates) {
-		if(!state.world.nation_get_is_substate(target))
-			return;
-		auto ruler = state.world.overlord_get_ruler(state.world.nation_get_overlord_as_subject(target));
-		if(can_use && !trigger::evaluate(state, can_use, trigger::to_generic(ruler), trigger::to_generic(n), trigger::to_generic(ruler))) {
-			return;
-		}
-	} else {
-		if(can_use && !trigger::evaluate(state, can_use, trigger::to_generic(target), trigger::to_generic(n), trigger::to_generic(target))) {
-			return;
-		}
-	}
-
-	auto allowed_countries = state.world.cb_type_get_allowed_countries(cb);
-	auto allowed_states = state.world.cb_type_get_allowed_states(cb);
-
-	if(!allowed_countries && allowed_states) {
-		bool any_allowed = false;
-		for(auto si : target_states) {
-			if(trigger::evaluate(state, allowed_states, trigger::to_generic(si), trigger::to_generic(n), trigger::to_generic(n))) {
-				assert(military::cb_conditions_satisfied(state, n, target, cb));
-				result.push_back(possible_cb{ target, dcon::nation_id{}, dcon::national_identity_id{}, state.world.state_instance_get_definition(si), cb });
-				return;
-			}
-		}
-		return;
-	} else if(allowed_substates) { // checking for whether the target is a substate is already done above
-		for(auto si : target_states) {
-			if(trigger::evaluate(state, allowed_substates, trigger::to_generic(si), trigger::to_generic(n), trigger::to_generic(n))) {
-				assert(military::cb_conditions_satisfied(state, n, target, cb));
-				result.push_back(possible_cb{ target, dcon::nation_id{}, dcon::national_identity_id{}, state.world.state_instance_get_definition(si), cb });
-				return;
-			}
-		}
-		return;
-	} else if(allowed_countries) {
-		bool liberate = (state.world.cb_type_get_type_bits(cb) & military::cb_flag::po_transfer_provinces) != 0;
-		for(auto other_nation : state.world.in_nation) {
-			if(other_nation != n) {
-				if(trigger::evaluate(state, allowed_countries, trigger::to_generic(target), trigger::to_generic(n),
-					trigger::to_generic(other_nation.id))) {
-					if(allowed_states) { // check whether any state within the target is valid for free / liberate
-						for(auto i = target_states.size(); i-- > 0;) {
-							auto si = target_states[i];
-							if(trigger::evaluate(state, allowed_states, trigger::to_generic(si), trigger::to_generic(n), trigger::to_generic(other_nation.id))) {
-								if(liberate) {
-									assert(military::cb_conditions_satisfied(state, n, target, cb));
-									result.push_back(possible_cb{ target, dcon::nation_id{}, other_nation.get_identity_from_identity_holder(), state.world.state_instance_get_definition(si), cb });
-									return;
-								} else {
-									assert(military::cb_conditions_satisfied(state, n, target, cb));
-									result.push_back(possible_cb{ target, other_nation, dcon::national_identity_id{}, state.world.state_instance_get_definition(si), cb });
-									return;
-								}
-							}
-						}
-					} else { // no allowed states trigger
-						if(liberate) {
-							assert(military::cb_conditions_satisfied(state, n, target, cb));
-							result.push_back(possible_cb{ target, dcon::nation_id{}, other_nation.get_identity_from_identity_holder(), dcon::state_definition_id{}, cb });
-							return;
-						} else {
-							assert(military::cb_conditions_satisfied(state, n, target, cb));
-							result.push_back(possible_cb{ target, other_nation, dcon::national_identity_id{}, dcon::state_definition_id{}, cb });
-							return;
-						}
-					}
-				}
-			}
-		}
-		return;
-	} else {
-		assert(military::cb_conditions_satisfied(state, n, target, cb));
-		result.push_back(possible_cb{ target, dcon::nation_id{}, dcon::national_identity_id{}, dcon::state_definition_id{}, cb });
-		return;
-	}
-}
-
-void sort_possible_justification_cbs(std::vector<possible_cb>& result, sys::state& state, dcon::nation_id n, dcon::nation_id target) {
-	result.clear();
-
-	static std::vector<dcon::state_instance_id> target_states;
-	state_target_list(target_states, state, n, target);
-
-	// AI can go after po_demand_state and po_annex
-	for(auto cb : state.world.in_cb_type) {
-		auto bits = state.world.cb_type_get_type_bits(cb);
-		if((bits & (military::cb_flag::po_demand_state | military::cb_flag::po_annex)) == 0)
-			continue;
-
-		if(!military::cb_conditions_satisfied(state, n, target, cb))
-			continue;
-		auto sl = state.world.nation_get_in_sphere_of(target);
-		if(sl == n)
-			continue;
-
-		place_instance_in_result(state, result, n, target, cb, target_states);
-	}
-
-	// Prioritize CB types
-	std::sort(result.begin(), result.end(), [&](possible_cb const& a, possible_cb const& b) {
-		if((state.world.nation_get_ai_rival(n) == a.target) != (state.world.nation_get_ai_rival(n) == b.target)) {
-			return state.world.nation_get_ai_rival(n) == a.target;
-		}
-
-		auto a_annexes = (state.world.cb_type_get_type_bits(a.cb) & military::cb_flag::po_annex) != 0;
-		auto b_annexes = (state.world.cb_type_get_type_bits(b.cb) & military::cb_flag::po_annex) != 0;
-		if(a_annexes != b_annexes)
-			return a_annexes;
-
-		auto a_land = (state.world.cb_type_get_type_bits(a.cb) & military::cb_flag::po_demand_state) != 0;
-		auto b_land = (state.world.cb_type_get_type_bits(b.cb) & military::cb_flag::po_demand_state) != 0;
-
-		if(a_land < b_land)
-			return a_land;
-
-		auto rel_a = state.world.get_diplomatic_relation_by_diplomatic_pair(n, a.target);
-		auto rel_b = state.world.get_diplomatic_relation_by_diplomatic_pair(n, b.target);
-		if(state.world.diplomatic_relation_get_value(rel_a) != state.world.diplomatic_relation_get_value(rel_b))
-			return state.world.diplomatic_relation_get_value(rel_a) < state.world.diplomatic_relation_get_value(rel_b);
-
-		if(a.cb != b.cb)
-			return a.cb.index() < b.cb.index();
-		if(a.target != b.target)
-			return a.target.index() < b.target.index();
-		if(a.secondary_nation != b.secondary_nation)
-			return a.secondary_nation.index() < b.secondary_nation.index();
-		if(a.associated_tag != b.associated_tag)
-			return a.associated_tag.index() < b.associated_tag.index();
-		return a.state_def.index() < b.state_def.index();
-	});
-}
-
-possible_cb pick_fabrication_type(sys::state& state, dcon::nation_id from, dcon::nation_id target) {
-	static std::vector<possible_cb> possibilities;
-
-	sort_possible_justification_cbs(possibilities, state, from, target);
-	// Uncivilized nations are more aggressive to westernize faster
-	float infamy_limit = state.world.nation_get_is_civilized(from) ? state.defines.badboy_limit / 2.f : state.defines.badboy_limit;
-
-	if(!possibilities.empty()) {
-		// AI only goes for top priority choice
-		auto possibility = possibilities[0];
-
-		// AI doesn't go over infamy limit;
-		if(military::cb_requires_selection_of_a_state(state, possibility.cb)) {
-			if(state.world.nation_get_infamy(from) + military::cb_infamy(state, possibility.cb, target, possibility.state_def) > infamy_limit)
-				return possible_cb{};
-		} else {
-			if(state.world.nation_get_infamy(from) + military::cb_infamy(state, possibility.cb, target) > infamy_limit)
-				return possible_cb{};
-		}
-
-		// Make sure no country higher than originator on rank is justifying on that target
-		for (auto n : state.nations_by_rank) {
-			if(n == from) {
-				break;
-			}
-
-			auto is_discovered = state.world.nation_get_constructing_cb_is_discovered(n);
-
-			if(!is_discovered) {
-				continue;
-			}
-
-			auto target_nation = state.world.nation_get_constructing_cb_target(n);
-			auto target_state = state.world.nation_get_constructing_cb_target_state(n);
-
-			if(possibility.target == target_nation && possibility.state_def == target_state) {
-				return possible_cb{};
-			}
-		}
-
-		return possibility;
-	} else {
-		return possible_cb{};
-	}
-}
-
-bool valid_construction_target(sys::state& state, dcon::nation_id from, dcon::nation_id target) {
-	// Copied from commands.cpp:can_fabricate_cb()
-	if(from == target)
-		return false;
-	if(state.world.nation_get_constructing_cb_type(from))
-		return false;
-	auto ol = state.world.nation_get_overlord_as_subject(from);
-	if(state.world.overlord_get_ruler(ol) && state.world.overlord_get_ruler(ol) != target)
-		return false;
-	if(state.world.nation_get_in_sphere_of(target) == from)
-		return false;
-	if(military::are_at_war(state, target, from))
-		return false;
-	if(military::has_truce_with(state, target, from))
-		return false;
-	auto sl = state.world.nation_get_in_sphere_of(target);
-	if(sl == from)
-		return false;
-	// Its easy to defeat a nation at war
-	if(state.world.nation_get_is_at_war(target)) {
-		if(estimate_strength(state, from) < estimate_strength(state, target) * 0.15f) {
-			return false;
-		}
-	} else {
-		if(estimate_strength(state, from) < estimate_strength(state, target) * 0.66f)
-			return false;
-	}
-	if(state.world.nation_get_owned_province_count(target) <= 0)
-		return false;
-
-	// if target is a landlocked nation or we are landlocked nation outselves
-	// and we are not neighbors
-	// we don't need any land of the nation
-	// because we cannot reliably benefit from it due to local control mechanic
-	if(nations::is_landlocked(state, target) || nations::is_landlocked(state, from)) {
-		if(!state.world.get_nation_adjacency_by_nation_adjacency_pair(from, target)) {
-			return false;
-		}
-	}
-
-	// Attacking people from other continents only if we have naval superiority
-	if(state.world.province_get_continent(state.world.nation_get_capital(from)) != state.world.province_get_continent(state.world.nation_get_capital(target))) {
-		// We must achieve naval superiority to even invade them
-		if(!naval_supremacy(state, from, target)) {
-			return false;
-		}
-		if(state.world.nation_get_capital_ship_score(from) < std::max(1.f, 1.5f * state.world.nation_get_capital_ship_score(target)))
-			return false;
-	}
-	return true;
-}
-
-void update_cb_fabrication(sys::state& state) {
-	for(auto nid : state.nations_by_rank) {
-		auto n = dcon::fatten(state.world, nid);
-
-		if(!n.get_is_player_controlled() && n.get_owned_province_count() > 0) {
-			if(n.get_is_at_war())
-				continue;
-			// Uncivilized nations are more aggressive to westernize faster
-			float infamy_limit = state.world.nation_get_is_civilized(n) ? state.defines.badboy_limit / 2.f : state.defines.badboy_limit;
-			if(n.get_infamy() > infamy_limit)
-				continue;
-			if(n.get_constructing_cb_type())
-				continue;
-			auto ol = n.get_overlord_as_subject().get_ruler().id;
-			if(n.get_ai_rival()
-				&& n.get_ai_rival().get_in_sphere_of() != n
-				&& (!ol || ol == n.get_ai_rival())
-				&& !military::are_at_war(state, n, n.get_ai_rival())
-				&& !military::can_use_cb_against(state, n, n.get_ai_rival())
-				&& !military::has_truce_with(state, n, n.get_ai_rival())) {
-
-				auto cb = pick_fabrication_type(state, n, n.get_ai_rival());
-				if(cb.cb) {
-					n.set_constructing_cb_target(n.get_ai_rival());
-					n.set_constructing_cb_type(cb.cb);
-					n.set_constructing_cb_target_state(cb.state_def);
-					continue;
-				}
-			}
-
-			static std::vector<std::pair<dcon::nation_id, float>> possible_targets;
-			possible_targets.clear();
-			float total_weight = 0.f;
-			for(auto i : state.world.in_nation) {
-				if(valid_construction_target(state, n, i)
-				&& !military::has_truce_with(state, n, i)) {
-
-					auto weight = 1.f / (province::direct_distance(
-						state,
-						state.world.nation_get_capital(n),
-						state.world.nation_get_capital(i)
-					) + 1.f);
-
-					// if nations are neighbors, then it's much easier to prepare a successful invasion:
-					if(state.world.get_nation_adjacency_by_nation_adjacency_pair(n, i)) {
-						weight *= 10.f;
-					}
-
-					// uncivs have higher priority
-					if(!i.get_is_civilized()) {
-						weight *= 2.f;
-					}
-
-					// nations with strong military have lower weight
-					weight = weight / (estimate_strength(state, i) + 1.f);
-
-					total_weight += weight;
-					possible_targets.push_back({ i.id, weight });
-				}
-			}
-			if(!possible_targets.empty()) {
-				auto random_value = uint32_t(rng::get_random(state, uint32_t(n.id.index())) >> 2);
-				auto limited_value = rng::reduce(random_value, 1000);
-				auto random_float = float(limited_value) / 1000.f * total_weight;
-				auto acc = 0.f;
-				dcon::nation_id target{ };
-				for(auto& weighted_nation : possible_targets) {
-					acc += weighted_nation.second;
-					if(acc > random_float) {
-						target = weighted_nation.first;
-						break;
-					}
-				}
-				if(target) {
-					if(auto pcb = pick_fabrication_type(state, n, target); pcb.cb) {
-						n.set_constructing_cb_target(target);
-						n.set_constructing_cb_type(pcb.cb);
-						n.set_constructing_cb_target_state(pcb.state_def);
-						continue;
-					}
-				}
-			}
-		}
-	}
-}
-
-bool will_join_war(sys::state& state, dcon::nation_id n, dcon::war_id w, bool as_attacker) {
-	/* Forbid war between uncivs against civs of other continent except if they are a sphereling or substate */
-	if(bool(state.defines.alice_unciv_civ_forbid_war) && !state.world.nation_get_is_civilized(n)
-	&& !state.world.nation_get_is_substate(n) && !state.world.nation_get_in_sphere_of(n)
-	&& !state.world.overlord_get_ruler(state.world.nation_get_overlord_as_subject(n))) {
-		auto pa = state.world.war_get_primary_attacker(w);
-		auto pd = state.world.war_get_primary_defender(w);
-		auto pa_cap = state.world.nation_get_capital(pa);
-		auto pd_cap = state.world.nation_get_capital(pd);
-		auto cap = state.world.nation_get_capital(n);
-		if(state.world.province_get_continent(pa_cap) != state.world.province_get_continent(cap))
-			return false;
-		if(state.world.province_get_continent(pd_cap) != state.world.province_get_continent(cap))
-			return false;
-	}
-
-	if(!as_attacker)
-		return true;
-	for(auto par : state.world.war_get_war_participant(w)) {
-		if(par.get_is_attacker() == !as_attacker) {
-			// Could use a CB against this nation?
-			if(military::can_use_cb_against(state, n, par.get_nation()))
-				return true;
-			// Eager to absolutely demolish our rival if possible
-			if(state.world.nation_get_ai_rival(n) == par.get_nation())
-				return true;
-		}
-	}
-	return false;
-}
-
-void place_instance_in_result_war(sys::state& state, std::vector<possible_cb>& result, dcon::nation_id n, dcon::nation_id target, dcon::war_id w, dcon::cb_type_id cb, std::vector<dcon::state_instance_id> const& target_states) {
-	auto can_use = state.world.cb_type_get_can_use(cb);
-	auto allowed_substates = state.world.cb_type_get_allowed_substate_regions(cb);
-	if(allowed_substates) {
-		if(!state.world.nation_get_is_substate(target))
-			return;
-		auto ruler = state.world.overlord_get_ruler(state.world.nation_get_overlord_as_subject(target));
-		if(can_use && !trigger::evaluate(state, can_use, trigger::to_generic(ruler), trigger::to_generic(n), trigger::to_generic(ruler))) {
-			return;
-		}
-	} else {
-		if(can_use && !trigger::evaluate(state, can_use, trigger::to_generic(target), trigger::to_generic(n), trigger::to_generic(target))) {
-			return;
-		}
-	}
-
-	auto allowed_countries = state.world.cb_type_get_allowed_countries(cb);
-	auto allowed_states = state.world.cb_type_get_allowed_states(cb);
-
-	if(!allowed_countries && allowed_states) {
-		for(auto si : target_states) {
-			if(trigger::evaluate(state, allowed_states, trigger::to_generic(si), trigger::to_generic(n), trigger::to_generic(n))) {
-				if(!military::war_goal_would_be_duplicate(state, n, w, target, cb, state.world.state_instance_get_definition(si), dcon::national_identity_id{}, dcon::nation_id{})) {
-					assert(military::cb_conditions_satisfied(state, n, target, cb));
-					result.push_back(possible_cb{ target, dcon::nation_id{}, dcon::national_identity_id{}, state.world.state_instance_get_definition(si), cb });
-					return;
-				}
-			}
-		}
-		return;
-	} else if(allowed_substates) { // checking for whether the target is a substate is already done above
-		for(auto si : target_states) {
-			if(trigger::evaluate(state, allowed_substates, trigger::to_generic(si), trigger::to_generic(n), trigger::to_generic(n))) {
-				if(!military::war_goal_would_be_duplicate(state, n, w, target, cb, state.world.state_instance_get_definition(si), dcon::national_identity_id{}, dcon::nation_id{})) {
-					assert(military::cb_conditions_satisfied(state, n, target, cb));
-					result.push_back(possible_cb{ target, dcon::nation_id{}, dcon::national_identity_id{}, state.world.state_instance_get_definition(si), cb });
-					return;
-				}
-			}
-		}
-		return;
-	} else if(allowed_countries) {
-		bool liberate = (state.world.cb_type_get_type_bits(cb) & military::cb_flag::po_transfer_provinces) != 0;
-		for(auto other_nation : state.world.in_nation) {
-			if(other_nation != n) {
-				if(trigger::evaluate(state, allowed_countries, trigger::to_generic(target), trigger::to_generic(n), trigger::to_generic(other_nation.id))) {
-					if(allowed_states) { // check whether any state within the target is valid for free / liberate
-						for(auto i = target_states.size(); i-- > 0;) {
-							auto si = target_states[i];
-							if(trigger::evaluate(state, allowed_states, trigger::to_generic(si), trigger::to_generic(n), trigger::to_generic(other_nation.id))) {
-								if(liberate) {
-									if(!military::war_goal_would_be_duplicate(state, n, w, target, cb, state.world.state_instance_get_definition(si), other_nation.get_identity_from_identity_holder(), dcon::nation_id{})) {
-										assert(military::cb_conditions_satisfied(state, n, target, cb));
-										result.push_back(possible_cb{ target, dcon::nation_id{}, other_nation.get_identity_from_identity_holder(), state.world.state_instance_get_definition(si), cb });
-										return;
-									}
-								} else {
-									if(!military::war_goal_would_be_duplicate(state, n, w, target, cb, state.world.state_instance_get_definition(si), dcon::national_identity_id{}, other_nation)) {
-										assert(military::cb_conditions_satisfied(state, n, target, cb));
-										result.push_back(possible_cb{ target, other_nation, dcon::national_identity_id{}, state.world.state_instance_get_definition(si), cb });
-										return;
-									}
-								}
-							}
-						}
-					} else { // no allowed states trigger
-						if(liberate) {
-							if(!military::war_goal_would_be_duplicate(state, n, w, target, cb, dcon::state_definition_id{}, other_nation.get_identity_from_identity_holder(), dcon::nation_id{})) {
-								assert(military::cb_conditions_satisfied(state, n, target, cb));
-								result.push_back(possible_cb{ target, dcon::nation_id{}, other_nation.get_identity_from_identity_holder(), dcon::state_definition_id{}, cb });
-								return;
-							}
-						} else {
-							if(!military::war_goal_would_be_duplicate(state, n, w, target, cb, dcon::state_definition_id{}, dcon::national_identity_id{}, other_nation)) {
-								assert(military::cb_conditions_satisfied(state, n, target, cb));
-								result.push_back(possible_cb{ target, other_nation, dcon::national_identity_id{}, dcon::state_definition_id{}, cb });
-								return;
-							}
-						}
-					}
-				}
-			}
-		}
-		return;
-	} else if(!military::war_goal_would_be_duplicate(state, n, w, target, cb, dcon::state_definition_id{}, dcon::national_identity_id{}, dcon::nation_id{})) {
-		assert(military::cb_conditions_satisfied(state, n, target, cb));
-		result.push_back(possible_cb{ target, dcon::nation_id{}, dcon::national_identity_id{}, dcon::state_definition_id{}, cb });
-		return;
-	}
-}
-
-void sort_available_cbs(std::vector<possible_cb>& result, sys::state& state, dcon::nation_id n, dcon::war_id w) {
-	result.clear();
-
-	auto is_attacker = military::get_role(state, w, n) == military::war_role::attacker;
-	for(auto par : state.world.war_get_war_participant(w)) {
-		if(par.get_is_attacker() != is_attacker) {
-			static std::vector<dcon::state_instance_id> target_states;
-			state_target_list(target_states, state, n, par.get_nation());
-			for(auto& cb : state.world.nation_get_available_cbs(n)) {
-				if(cb.target == par.get_nation())
-					place_instance_in_result_war(state, result, n, par.get_nation(), w, cb.cb_type, target_states);
-			}
-			for(auto cb : state.world.in_cb_type) {
-				if((cb.get_type_bits() & military::cb_flag::always) != 0) {
-					place_instance_in_result_war(state, result, n, par.get_nation(), w, cb, target_states);
-				}
-			}
-		}
-	}
-
-	std::sort(result.begin(), result.end(), [&](possible_cb const& a, possible_cb const& b) {
-		if((state.world.nation_get_ai_rival(n) == a.target) != (state.world.nation_get_ai_rival(n) == b.target)) {
-			return state.world.nation_get_ai_rival(n) == a.target;
-		}
-
-		auto a_annexes = (state.world.cb_type_get_type_bits(a.cb) & military::cb_flag::po_annex) != 0;
-		auto b_annexes = (state.world.cb_type_get_type_bits(b.cb) & military::cb_flag::po_annex) != 0;
-		if(a_annexes != b_annexes)
-			return a_annexes;
-
-		auto a_land = (state.world.cb_type_get_type_bits(a.cb) & military::cb_flag::po_demand_state) != 0;
-		auto b_land = (state.world.cb_type_get_type_bits(b.cb) & military::cb_flag::po_demand_state) != 0;
-
-		if(a_land < b_land)
-			return a_land;
-
-		auto rel_a = state.world.get_diplomatic_relation_by_diplomatic_pair(n, a.target);
-		auto rel_b = state.world.get_diplomatic_relation_by_diplomatic_pair(n, b.target);
-		if(state.world.diplomatic_relation_get_value(rel_a) != state.world.diplomatic_relation_get_value(rel_b))
-			return state.world.diplomatic_relation_get_value(rel_a) < state.world.diplomatic_relation_get_value(rel_b);
-
-		if(a.cb != b.cb)
-			return a.cb.index() < b.cb.index();
-		if(a.target != b.target)
-			return a.target.index() < b.target.index();
-		if(a.secondary_nation != b.secondary_nation)
-			return a.secondary_nation.index() < b.secondary_nation.index();
-		if(a.associated_tag != b.associated_tag)
-			return a.associated_tag.index() < b.associated_tag.index();
-		return a.state_def.index() < b.state_def.index();
-	});
-}
-
-void sort_available_declaration_cbs(std::vector<possible_cb>& result, sys::state& state, dcon::nation_id n, dcon::nation_id target) {
-	result.clear();
-
-	static std::vector<dcon::state_instance_id> target_states;
-	state_target_list(target_states, state, n, target);
-
-	// Justified CBs
-	auto other_cbs = state.world.nation_get_available_cbs(n);
-	for(auto& cb : other_cbs) {
-		if(cb.target == target)
-			place_instance_in_result(state, result, n, target, cb.cb_type, target_states);
-	}
-	// Always available CBs (incl acquire_core_state and annex_core_country)
-	for(auto cb : state.world.in_cb_type) {
-		if((cb.get_type_bits() & military::cb_flag::always) != 0) {
-			place_instance_in_result(state, result, n, target, cb, target_states);
-		}
-	}
-
-	// Prioritize CB types
-	std::sort(result.begin(), result.end(), [&](possible_cb const& a, possible_cb const& b) {
-		if((state.world.nation_get_ai_rival(n) == a.target) != (state.world.nation_get_ai_rival(n) == b.target)) {
-			return state.world.nation_get_ai_rival(n) == a.target;
-		}
-
-		auto a_annexes = (state.world.cb_type_get_type_bits(a.cb) & military::cb_flag::po_annex) != 0;
-		auto b_annexes = (state.world.cb_type_get_type_bits(b.cb) & military::cb_flag::po_annex) != 0;
-		if(a_annexes != b_annexes)
-			return a_annexes;
-
-		auto a_land = (state.world.cb_type_get_type_bits(a.cb) & military::cb_flag::po_demand_state) != 0;
-		auto b_land = (state.world.cb_type_get_type_bits(b.cb) & military::cb_flag::po_demand_state) != 0;
-
-		if(a_land < b_land)
-			return a_land;
-
-		auto rel_a = state.world.get_diplomatic_relation_by_diplomatic_pair(n, a.target);
-		auto rel_b = state.world.get_diplomatic_relation_by_diplomatic_pair(n, b.target);
-		if(state.world.diplomatic_relation_get_value(rel_a) != state.world.diplomatic_relation_get_value(rel_b))
-			return state.world.diplomatic_relation_get_value(rel_a) < state.world.diplomatic_relation_get_value(rel_b);
-
-		if(a.cb != b.cb)
-			return a.cb.index() < b.cb.index();
-		if(a.target != b.target)
-			return a.target.index() < b.target.index();
-		if(a.secondary_nation != b.secondary_nation)
-			return a.secondary_nation.index() < b.secondary_nation.index();
-		if(a.associated_tag != b.associated_tag)
-			return a.associated_tag.index() < b.associated_tag.index();
-		return a.state_def.index() < b.state_def.index();
-	});
-}
-
-void add_free_ai_cbs_to_war(sys::state& state, dcon::nation_id n, dcon::war_id w) {
-	bool is_attacker = military::is_attacker(state, w, n);
-	if(!is_attacker && military::defenders_have_status_quo_wargoal(state, w))
-		return;
-	if(is_attacker && military::attackers_have_status_quo_wargoal(state, w))
-		return;
-
-	bool added = false;
-	do {
-		added = false;
-		static std::vector<possible_cb> potential;
-		sort_available_cbs(potential, state, n, w);
-		for(auto& p : potential) {
-			if(!military::war_goal_would_be_duplicate(state, n, w, p.target, p.cb, p.state_def, p.associated_tag, p.secondary_nation)) {
-				military::add_wargoal(state, w, n, p.target, p.cb, p.state_def, p.associated_tag, p.secondary_nation);
-				nations::adjust_relationship(state, n, p.target, state.defines.addwargoal_relation_on_accept);
-				added = true;
-			}
-		}
-	} while(added);
-
-}
-
-dcon::cb_type_id pick_wargoal_extra_cb_type(sys::state& state, dcon::nation_id from, dcon::nation_id target, bool is_great) {
-	static std::vector<dcon::cb_type_id> possibilities;
-	possibilities.clear();
-
-	auto free_infamy = state.defines.badboy_limit - state.world.nation_get_infamy(from);
-
-	auto multiplier = is_great ? state.defines.gw_justify_cb_badboy_impact : 1.f;
-
-	for(auto c : state.world.in_cb_type) {
-		auto bits = state.world.cb_type_get_type_bits(c);
-		if((bits & (military::cb_flag::always | military::cb_flag::is_not_constructing_cb)) != 0)
-			continue;
-		if((bits & (military::cb_flag::po_demand_state | military::cb_flag::po_annex)) == 0)
-			continue;
-		if(military::cb_infamy(state, c, target) * multiplier > free_infamy)
-			continue;
-		if(!military::cb_conditions_satisfied(state, from, target, c))
-			continue;
-		if(!valid_construction_target(state, from, target))
-			continue;
-
-		possibilities.push_back(c);
-	}
-
-	if(!possibilities.empty()) {
-		return possibilities[rng::reduce(uint32_t(rng::get_random(state, uint32_t((from.index() << 3) ^ target.index()))), uint32_t(possibilities.size()))];
-	} else {
-		return dcon::cb_type_id{};
-	}
-}
-
-dcon::nation_id pick_wargoal_target(sys::state& state, dcon::nation_id from, dcon::war_id w, bool is_attacker) {
-
-	if(is_attacker && military::get_role(state, w, state.world.nation_get_ai_rival(from)) == military::war_role::defender)
-		return state.world.nation_get_ai_rival(from);
-	if(!is_attacker && military::get_role(state, w, state.world.nation_get_ai_rival(from)) == military::war_role::attacker)
-		return state.world.nation_get_ai_rival(from);
-
-	static std::vector<dcon::nation_id> possibilities;
-	possibilities.clear();
-
-	for(auto par : state.world.war_get_war_participant(w)) {
-		if(par.get_is_attacker() != is_attacker) {
-			if(state.world.get_nation_adjacency_by_nation_adjacency_pair(from, par.get_nation()))
-				possibilities.push_back(par.get_nation().id);
-		}
-	}
-	if(possibilities.empty()) {
-		for(auto par : state.world.war_get_war_participant(w)) {
-			if(par.get_is_attacker() != is_attacker) {
-				if(nations::is_great_power(state, par.get_nation()))
-					possibilities.push_back(par.get_nation().id);
-			}
-		}
-	}
-	if(!possibilities.empty()) {
-		return possibilities[rng::reduce(uint32_t(rng::get_random(state, uint32_t(from.index() ^ 3))), uint32_t(possibilities.size()))];
-	} else {
-		return dcon::nation_id{};
-	}
-}
-
-void add_wargoal_to_war(sys::state& state, dcon::nation_id n, dcon::war_id w) {
-	auto rval = rng::get_random(state, n.index() ^ w.index() << 2);
-	if((rval & 1) == 0)
-		return;
-
-	if(n == state.world.war_get_primary_attacker(w) || n == state.world.war_get_primary_defender(w)) {
-		if(((rval >> 1) & 1) == 0) {
-			add_free_ai_cbs_to_war(state, n, w);
-		}
-	}
-
-	auto totalpop = state.world.nation_get_demographics(n, demographics::total);
-	auto jingoism_perc = totalpop > 0 ? state.world.nation_get_demographics(n, demographics::to_key(state, state.culture_definitions.jingoism)) / totalpop : 0.0f;
-
-	if(jingoism_perc < state.defines.wargoal_jingoism_requirement * state.defines.gw_wargoal_jingoism_requirement_mod)
-		return;
-
-	bool attacker = military::get_role(state, w, n) == military::war_role::attacker;
-	auto spare_ws = attacker ? (military::primary_warscore(state, w) - military::attacker_peace_cost(state, w)) : (-military::primary_warscore(state, w) - military::defender_peace_cost(state, w));
-	if(spare_ws < 1.0f)
-		return;
-
-	auto target = pick_wargoal_target(state, n, w, attacker);
-	if(!target)
-		return;
-
-	auto cb = pick_wargoal_extra_cb_type(state, n, target, state.world.war_get_is_great(w));
-	if(!cb)
-		return;
-
-	auto multiplier = state.world.war_get_is_great(w) ? state.defines.gw_justify_cb_badboy_impact : 1.f;
-
-	static std::vector<dcon::state_instance_id> target_states;
-	state_target_list(target_states, state, n, target);
-	static std::vector<possible_cb> result;
-	result.clear();
-	place_instance_in_result_war(state, result, n, target, w, cb, target_states);
-	if(!result.empty() && result[0].target) {
-		military::add_wargoal(state, w, n, target, cb, result[0].state_def, result[0].associated_tag, result[0].secondary_nation);
-		nations::adjust_relationship(state, n, target, state.defines.addwargoal_relation_on_accept);
-		state.world.nation_get_infamy(n) += military::cb_infamy(state, cb, target) * multiplier;
-	}
-}
-
-void add_wargoals(sys::state& state) {
-	for(auto w : state.world.in_war) {
-		for(auto par : w.get_war_participant()) {
-			if(par.get_nation().get_is_player_controlled() == false) {
-				add_wargoal_to_war(state, par.get_nation(), w);
-			}
-		}
-	}
-}
-
-// Every single owned core of the country is occupied
-bool has_cores_occupied(sys::state& state, dcon::nation_id n) {
-	if(bool(state.defines.alice_surrender_on_cores_lost)) {
-		auto i = state.world.nation_get_identity_from_identity_holder(n);
-		auto cores = state.world.national_identity_get_core(i);
-		bool has_owned_cores = false;
-		for(auto c : cores) {
-			if(c.get_province().get_nation_from_province_ownership() == n) {
-				has_owned_cores = true;
-			} if(c.get_province().get_nation_from_province_control() == n) {
-				return false;
-			}
-		}
-		auto pc = state.world.nation_get_province_control(n); //no cores or some cores are occupied but some are not
-		return has_owned_cores || pc.begin() == pc.end(); //controls anything?
-	}
-	return false;
-}
-
-void make_peace_offers(sys::state& state) {
-	auto send_offer_up_to = [&](dcon::nation_id from, dcon::nation_id to, dcon::war_id w, bool attacker, int32_t score_max, bool concession) {
-		if(auto off = state.world.nation_get_peace_offer_from_pending_peace_offer(from); off) {
-			if(state.world.peace_offer_get_is_crisis_offer(off) == true || state.world.peace_offer_get_war_from_war_settlement(off))
-				return; // offer in flight
-			state.world.delete_peace_offer(off); // else -- offer has been already resolved and was just pending gc
-		}
-
-		assert(command::can_start_peace_offer(state, from, to, w, concession));
-		command::execute_start_peace_offer(state, from, to, w, concession);
-		auto pending = state.world.nation_get_peace_offer_from_pending_peace_offer(from);
-		if(!pending)
-			return;
-
-		score_max = std::min(score_max, 100);
-		int32_t current_value = 0;
-		for(auto wg : state.world.war_get_wargoals_attached(w)) {
-			if((military::is_attacker(state, w, wg.get_wargoal().get_added_by()) == attacker) == !concession) {
-				auto goal_cost = military::peace_cost(state, w, wg.get_wargoal().get_type(), wg.get_wargoal().get_added_by(), wg.get_wargoal().get_target_nation(), wg.get_wargoal().get_secondary_nation(), wg.get_wargoal().get_associated_state(), wg.get_wargoal().get_associated_tag());
-				if(current_value + goal_cost <= score_max) {
-					current_value += goal_cost;
-					state.world.force_create_peace_offer_item(pending, wg.get_wargoal().id);
-				}
-			}
-		}
-
-		assert(command::can_send_peace_offer(state, from));
-		command::execute_send_peace_offer(state, from);
-	};
-
-	for(auto w : state.world.in_war) {
-		if((w.get_primary_attacker().get_is_player_controlled() == false || w.get_primary_defender().get_is_player_controlled() == false)
-		&& w.get_primary_attacker().get_owned_province_count() > 0
-		&& w.get_primary_defender().get_owned_province_count() > 0) {
-			//postpone until military gc does magic
-			if(military::get_role(state, w, w.get_primary_attacker()) != military::war_role::attacker)
-				continue;
-			if(military::get_role(state, w, w.get_primary_defender()) != military::war_role::defender)
-				continue;
-
-			auto overall_score = military::primary_warscore(state, w);
-			if(overall_score >= 0) { // attacker winning
-				bool defender_surrender = has_cores_occupied(state, w.get_primary_defender());
-				auto total_po_cost = military::attacker_peace_cost(state, w);
-				if(w.get_primary_attacker().get_is_player_controlled() == false) { // attacker makes offer
-					if(defender_surrender || (overall_score >= 100 || (overall_score >= 50 && overall_score >= total_po_cost * 2))) {
-						send_offer_up_to(w.get_primary_attacker(), w.get_primary_defender(), w, true, int32_t(overall_score), false);
-						continue;
-					}
-					if(w.get_primary_defender().get_is_player_controlled() == false) {
-						auto war_duration = state.current_date.value - state.world.war_get_start_date(w).value;
-						if(war_duration >= 365) {
-							float willingness_factor = float(war_duration - 365) * 10.f / 365.0f;
-							if(defender_surrender || (overall_score > (total_po_cost - willingness_factor) && (-overall_score / 2 + total_po_cost - willingness_factor) < 0)) {
-								send_offer_up_to(w.get_primary_attacker(), w.get_primary_defender(), w, true, int32_t(total_po_cost), false);
-								continue;
-							}
-						}
-					}
-				} else if(w.get_primary_defender().get_is_player_controlled() == false) { // defender may surrender
-					if(defender_surrender || (overall_score >= 100 || (overall_score >= 50 && overall_score >= total_po_cost * 2))) {
-						send_offer_up_to(w.get_primary_defender(), w.get_primary_attacker(), w, false, int32_t(overall_score), true);
-						continue;
-					}
-				}
-			} else {
-				bool attacker_surrender = has_cores_occupied(state, w.get_primary_attacker());
-				auto total_po_cost = military::defender_peace_cost(state, w);
-				if(w.get_primary_defender().get_is_player_controlled() == false) { // defender makes offer
-					if(attacker_surrender || (overall_score <= -100 || (overall_score <= -50 && overall_score <= -total_po_cost * 2))) {
-						send_offer_up_to(w.get_primary_defender(), w.get_primary_attacker(), w, false, int32_t(-overall_score), false);
-						continue;
-					}
-					if(w.get_primary_attacker().get_is_player_controlled() == false) {
-						auto war_duration = state.current_date.value - state.world.war_get_start_date(w).value;
-						if(war_duration >= 365) {
-							float willingness_factor = float(war_duration - 365) * 10.f / 365.0f;
-							if(attacker_surrender  || (-overall_score > (total_po_cost - willingness_factor) && (overall_score / 2 + total_po_cost - willingness_factor) < 0)) {
-								send_offer_up_to(w.get_primary_defender(), w.get_primary_attacker(), w, false, int32_t(total_po_cost), false);
-								continue;
-							}
-						}
-					}
-				} else if(w.get_primary_attacker().get_is_player_controlled() == false) { // attacker may surrender
-					if(attacker_surrender || (overall_score <= -100 || (overall_score <= -50 && overall_score <= -total_po_cost * 2))) {
-						send_offer_up_to(w.get_primary_attacker(), w.get_primary_defender(), w, true, int32_t(-overall_score), true);
-						continue;
-					}
-				}
-			}
-		}
-	}
-}
-
-bool will_accept_peace_offer_value(sys::state& state,
-	dcon::nation_id n, // TO country reviews peace offer
-	dcon::nation_id from, // FROM country sends peace offer
-	dcon::nation_id prime_attacker, dcon::nation_id prime_defender,
-	float primary_warscore,
-	float scoreagainst_me, // Score against TO country
-	bool offer_from_attacker,
-	bool concession, // False if FROM country demands from TO. True if FROM submits goals to TO.
-	int32_t overall_po_value, // Score value of the proposed deal
-	int32_t my_po_target, // How much warscore worth of wargoals does TO country have against FROM
-	int32_t target_personal_po_value, int32_t potential_peace_score_against,
-	int32_t my_side_against_target, int32_t my_side_peace_cost,
-	int32_t war_duration, // Length of the war in days
-	bool contains_sq // Does peace offer contain status quo
-) {
-	bool is_attacking = !offer_from_attacker;
-
-	auto overall_score = primary_warscore;
-	if(concession && overall_score <= -50.0f) {
-		return true;
-	}
-
-	if(!concession) {
-		overall_po_value = -overall_po_value;
-	}
-	if(overall_po_value < -100)
-		return false;
-
-	auto personal_score_saved = target_personal_po_value - potential_peace_score_against;
-	auto war_exhaustion = state.world.nation_get_war_exhaustion(n); // War exhaustion between 0 and 100
-	// Since we have functional blockades now, it's reasonable to account for war_exhaustion in AI willingness to peace out.
-	float willingness_factor = float(war_duration - 365) * 10.f / 365.0f + war_exhaustion;
-
-	// War-ending peace from primary participant to primary participant (concession & demand)
-	if((prime_attacker == n || prime_defender == n) && (prime_attacker == from || prime_defender == from)) {
-		if((overall_score <= -50 || war_exhaustion > state.defines.alice_ai_war_exhaustion_readiness_limit) && overall_score <= overall_po_value * 2)
-			return true;
-
-		// Forced peace when 100 warscore
-		if(overall_score <= -100 and overall_po_value <= 100)
-			return true;
-
-		if(concession && my_side_peace_cost <= overall_po_value)
-			return true; // offer contains everything
-		if(war_duration < 365) {
-			return false;
-		}
-		
-		if(overall_score >= 0) {
-			if(concession && ((overall_score * 2 - overall_po_value - willingness_factor) < 0))
-				return true;
-		} else {
-			if((overall_score - willingness_factor) <= overall_po_value && (overall_score / 2 - overall_po_value - willingness_factor) < 0)
-				return true;
-		}
-	}
-	// Peace offer from secondary participant to primary participant (concession)
-	else if((prime_attacker == n || prime_defender == n) && concession) {
-		if(scoreagainst_me > 50 || war_exhaustion > state.defines.alice_ai_war_exhaustion_readiness_limit)
-			return true;
-
-		if(overall_score < 0.0f) { // we are losing
-			if(my_side_against_target - scoreagainst_me <= overall_po_value + personal_score_saved)
-				return true;
-		} else {
-			if(my_side_against_target <= overall_po_value)
-				return true;
-		}
-
-	}
-	// Peace offer to secondary participant (concession & demand)
-	else {
-		if(contains_sq)
-			return false;
-
-		if((scoreagainst_me > 50 || war_exhaustion > state.defines.alice_ai_war_exhaustion_readiness_limit) && scoreagainst_me > -overall_po_value * 2)
-			return true;
-
-		// Forced peace when 100 warscore
-		if(scoreagainst_me >= 100 and overall_po_value <= 100)
-			return true;
-
-		if(overall_score < 0.0f) { // we are losing
-			if(personal_score_saved > 0 && scoreagainst_me + personal_score_saved - my_po_target >= -overall_po_value)
-				return true;
-
-		} else { // we are winning
-			if(my_po_target > 0 && my_po_target >= overall_po_value)
-				return true;
-		}
-	}
-
-	//will accept anything
-	if(has_cores_occupied(state, n))
-		return true;
-	return false;
-}
-
-bool will_accept_peace_offer(sys::state& state, dcon::nation_id n, dcon::nation_id from, dcon::peace_offer_id p) {
-	auto w = state.world.peace_offer_get_war_from_war_settlement(p);
-	auto prime_attacker = state.world.war_get_primary_attacker(w);
-	auto prime_defender = state.world.war_get_primary_defender(w);
-	bool is_attacking = military::is_attacker(state, w, n);
-	bool contains_sq = false;
-
-	auto overall_score = military::primary_warscore(state, w);
-	if(!is_attacking)
-		overall_score = -overall_score;
-
-	auto concession = state.world.peace_offer_get_is_concession(p);
-
-	if(concession && overall_score <= -50.0f) {
-		return true;
-	}
-
-	int32_t overall_po_value = 0;
-	int32_t personal_po_value = 0;
-	int32_t my_po_target = 0;
-	for(auto wg : state.world.peace_offer_get_peace_offer_item(p)) {
-		auto wg_value = military::peace_cost(state, w, wg.get_wargoal().get_type(), wg.get_wargoal().get_added_by(), wg.get_wargoal().get_target_nation(), wg.get_wargoal().get_secondary_nation(), wg.get_wargoal().get_associated_state(), wg.get_wargoal().get_associated_tag());
-		overall_po_value += wg_value;
-
-		if((wg.get_wargoal().get_type().get_type_bits() & military::cb_flag::po_status_quo) != 0)
-			contains_sq = true;
-
-		if(wg.get_wargoal().get_target_nation() == n) {
-			personal_po_value += wg_value;
-		}
-	}
-	if(!concession) {
-		overall_po_value = -overall_po_value;
-	}
-	if(overall_po_value < -100)
-		return false;
-
-
-	int32_t potential_peace_score_against = 0;
-	for(auto wg : state.world.war_get_wargoals_attached(w)) {
-		if(wg.get_wargoal().get_target_nation() == n || wg.get_wargoal().get_added_by() == n) {
-			auto wg_value = military::peace_cost(state, w, wg.get_wargoal().get_type(), wg.get_wargoal().get_added_by(), n, wg.get_wargoal().get_secondary_nation(), wg.get_wargoal().get_associated_state(), wg.get_wargoal().get_associated_tag());
-
-			if(wg.get_wargoal().get_target_nation() == n && (wg.get_wargoal().get_added_by() == from || from == prime_attacker || from == prime_defender)) {
-				potential_peace_score_against += wg_value;
-			}
-			if(wg.get_wargoal().get_added_by() == n && (wg.get_wargoal().get_target_nation() == from || from == prime_attacker || from == prime_defender)) {
-				my_po_target += wg_value;
-			}
-		}
-	}
-	auto personal_score_saved = personal_po_value - potential_peace_score_against;
-
-	auto war_duration = state.current_date.value - state.world.war_get_start_date(w).value;
-	auto war_exhaustion = state.world.nation_get_war_exhaustion(n); // War exhaustion between 0 and 100
-	// Since we have functional blockades now, it's reasonable to account for war_exhaustion in AI willingness to peace out.
-	float willingness_factor = float(war_duration - 365) * 10.f / 365.0f + war_exhaustion;
-
-	// War-ending peace from primary participant to primary participant (concession & demand)
-	if((prime_attacker == n || prime_defender == n) && (prime_attacker == from || prime_defender == from)) {
-		if((overall_score <= -50 || war_exhaustion > state.defines.alice_ai_war_exhaustion_readiness_limit) && overall_score <= overall_po_value * 2)
-			return true;
-
-		if(concession && (is_attacking ? military::attacker_peace_cost(state, w) : military::defender_peace_cost(state, w)) <= overall_po_value)
-			return true; // offer contains everything
-		if(war_duration < 365) {
-			return false;
-		}
-		if(overall_score >= 0) {
-			if(concession && ((overall_score * 2 - overall_po_value - willingness_factor) < 0))
-				return true;
-		} else {
-			if((overall_score - willingness_factor) <= overall_po_value && (overall_score / 2 - overall_po_value - willingness_factor) < 0)
-				return true;
-		}
-
-	}
-	// Peace offer from secondary participant to primary participant (concession)
-	else if((prime_attacker == n || prime_defender == n) && concession) {
-		auto scoreagainst_me = military::directed_warscore(state, w, from, n);
-
-		if(scoreagainst_me > 50 || war_exhaustion > state.defines.alice_ai_war_exhaustion_readiness_limit)
-			return true;
-
-		int32_t my_side_against_target = 0;
-		for(auto wg : state.world.war_get_wargoals_attached(w)) {
-			if(wg.get_wargoal().get_target_nation() == from) {
-				auto wg_value = military::peace_cost(state, w, wg.get_wargoal().get_type(), wg.get_wargoal().get_added_by(), n, wg.get_wargoal().get_secondary_nation(), wg.get_wargoal().get_associated_state(), wg.get_wargoal().get_associated_tag());
-
-				my_side_against_target += wg_value;
-			}
-		}
-
-		if(overall_score < 0.0f) { // we are losing
-			if(my_side_against_target - scoreagainst_me <= overall_po_value + personal_score_saved)
-				return true;
-		} else {
-			if(my_side_against_target <= overall_po_value)
-				return true;
-		}
-	}
-	// Peace offer to secondary participant (concession & demand)
-	else {
-		if(contains_sq)
-			return false;
-
-		auto scoreagainst_me = military::directed_warscore(state, w, from, n);
-		if((scoreagainst_me > 50 || war_exhaustion > state.defines.alice_ai_war_exhaustion_readiness_limit) && scoreagainst_me > -overall_po_value * 2)
-			return true;
-
-		if(overall_score < 0.0f) { // we are losing
-			if(personal_score_saved > 0 && scoreagainst_me + personal_score_saved - my_po_target >= -overall_po_value)
-				return true;
-
-		} else { // we are winning
-			if(my_po_target > 0 && my_po_target >= overall_po_value)
-				return true;
-		}
-	}
-
-	//will accept anything
-	if(has_cores_occupied(state, n))
-		return true;
-	return false;
-}
-
-
-void make_war_decs(sys::state& state) {
-	auto targets = ve::vectorizable_buffer<dcon::nation_id, dcon::nation_id>(state.world.nation_size());
-	concurrency::parallel_for(uint32_t(0), state.world.nation_size(), [&](uint32_t i) {
-		dcon::nation_id n{ dcon::nation_id::value_base_t(i) };
-		if(state.world.nation_get_owned_province_count(n) == 0)
-			return;
-		if(state.world.nation_get_is_at_war(n))
-			return;
-		if(state.world.nation_get_is_player_controlled(n))
-			return;
-		if(state.world.nation_get_military_score(n) == 0)
-			return;
-		// Subjects don't declare wars without a define set
-		if(auto ol = state.world.nation_get_overlord_as_subject(n); state.world.overlord_get_ruler(ol) && state.defines.alice_allow_subjects_declare_wars < 1)
-			return;
-		// AI shouldn't declare wars if it has high war exhaustion that would make it want to peace out immediately
-		if(state.world.nation_get_war_exhaustion(n) > state.defines.alice_ai_war_exhaustion_readiness_limit)
-			return;
-		auto base_strength = estimate_strength(state, n);
-		float best_difference = 2.0f;
-		//Great powers should look for non-neighbor nations to use their existing wargoals on; helpful for forcing unification/repay debts wars to happen
-		if(nations::is_great_power(state, n)) {
-			for(auto target : state.world.in_nation) {
-				auto real_target = target.get_overlord_as_subject().get_ruler() ? target.get_overlord_as_subject().get_ruler() : target;
-				if(target == n || real_target == n)
-					continue;
-				if(state.world.nation_get_owned_province_count(real_target) == 0)
-					continue;
-				if(nations::are_allied(state, n, real_target))
-					continue;
-				if(target.get_in_sphere_of() == n)
-					continue;
-				if(military::has_truce_with(state, n, real_target))
-					continue;
-				if(!military::can_use_cb_against(state, n, target))
-					continue;
-				//If it neighbors one of our spheres and we can pathfind to each other's capitals, we don't need naval supremacy to reach this nation
-				//Generally here to help Prussia realize it doesn't need a navy to attack Denmark
-				for(auto adj : state.world.nation_get_nation_adjacency(target)) {
-					auto other = adj.get_connected_nations(0) != n ? adj.get_connected_nations(0) : adj.get_connected_nations(1);
-					auto neighbor = other;
-					if(neighbor.get_in_sphere_of() == n){
-						auto path = province::make_safe_land_path(state, state.world.nation_get_capital(n), state.world.nation_get_capital(neighbor), n);
-						if(path.empty()) {
-							continue;
-						}
-						auto str_difference = base_strength + estimate_additional_offensive_strength(state, n, real_target) - estimate_defensive_strength(state, real_target);
-						if(str_difference > best_difference) {
-							best_difference = str_difference;
-							targets.set(n, target.id);
-							break;
-						}
-					}
-				}
-				if(!state.world.get_nation_adjacency_by_nation_adjacency_pair(n, target) && !naval_supremacy(state, n, target))
-					continue;
-				auto str_difference = base_strength + estimate_additional_offensive_strength(state, n, real_target) - estimate_defensive_strength(state, real_target);
-				if(str_difference > best_difference) {
-					best_difference = str_difference;
-					targets.set(n, target.id);
-				}
-			}
-		}
-		for(auto adj : state.world.nation_get_nation_adjacency(n)) {
-			auto other = adj.get_connected_nations(0) != n ? adj.get_connected_nations(0) : adj.get_connected_nations(1);
-			auto real_target = other.get_overlord_as_subject().get_ruler() ? other.get_overlord_as_subject().get_ruler() : other;
-			if(real_target == n)
-				continue;
-			if(nations::are_allied(state, n, real_target) || nations::are_allied(state, n, other))
-				continue;
-			if(real_target.get_in_sphere_of() == n)
-				continue;
-			if(state.world.nation_get_in_sphere_of(other) == n)
-				continue;
-			if(military::has_truce_with(state, n, other) || military::has_truce_with(state, n, real_target))
-				continue;
-			if(!military::can_use_cb_against(state, n, other))
-				continue;
-			if(!state.world.get_nation_adjacency_by_nation_adjacency_pair(n, other) && !naval_supremacy(state, n, other))
-				continue;
-			auto str_difference = base_strength + estimate_additional_offensive_strength(state, n, real_target) - estimate_defensive_strength(state, real_target);
-			if(str_difference > best_difference) {
-				best_difference = str_difference;
-				targets.set(n, other.id);
-			}
-		}
-		if(state.world.nation_get_central_ports(n) > 0) {
-			// try some random coastal nations
-			for(uint32_t j = 0; j < 6; ++j) {
-				auto rvalue = rng::get_random(state, uint32_t((n.index() << 3) + j));
-				auto reduced_value = rng::reduce(uint32_t(rvalue), state.world.nation_size());
-				dcon::nation_id other{ dcon::nation_id::value_base_t(reduced_value) };
-				auto real_target = fatten(state.world, other).get_overlord_as_subject().get_ruler() ? fatten(state.world, other).get_overlord_as_subject().get_ruler() : fatten(state.world, other);
-				if(other == n || real_target == n)
-					continue;
-				if(state.world.nation_get_owned_province_count(other) == 0 || state.world.nation_get_owned_province_count(real_target) == 0)
-					continue;
-				if(state.world.nation_get_central_ports(other) == 0 || state.world.nation_get_central_ports(real_target) == 0)
-					continue;
-				if(nations::are_allied(state, n, real_target) || nations::are_allied(state, n, other))
-					continue;
-				if(real_target.get_in_sphere_of() == n)
-					continue;
-				if(state.world.nation_get_in_sphere_of(other) == n)
-					continue;
-				if(military::has_truce_with(state, n, other) || military::has_truce_with(state, n, real_target))
-					continue;
-				if(!military::can_use_cb_against(state, n, other))
-					continue;
-				if(!state.world.get_nation_adjacency_by_nation_adjacency_pair(n, other) && !naval_supremacy(state, n, other))
-					continue;
-				auto str_difference = base_strength + estimate_additional_offensive_strength(state, n, real_target) - estimate_defensive_strength(state, real_target);
-				if(str_difference > best_difference) {
-					best_difference = str_difference;
-					targets.set(n, other);
-				}
-			}
-		}
-	});
-	for(auto n : state.world.in_nation) {
-		if(n.get_is_at_war() == false && targets.get(n)) {
-			static std::vector<possible_cb> potential;
-			sort_available_declaration_cbs(potential, state, n, targets.get(n));
-			if(!potential.empty()) {
-				assert(command::can_declare_war(state, n, targets.get(n), potential[0].cb, potential[0].state_def, potential[0].associated_tag, potential[0].secondary_nation));
-				command::execute_declare_war(state, n, targets.get(n), potential[0].cb, potential[0].state_def, potential[0].associated_tag, potential[0].secondary_nation, true, false);
-			}
-		}
-	}
-}
-
-void update_budget(sys::state& state) {
-	concurrency::parallel_for(uint32_t(0), state.world.nation_size(), [&](uint32_t i) {
-		dcon::nation_id nid{ dcon::nation_id::value_base_t(i) };
-		auto n = fatten(state.world, nid);
-		if(n.get_is_player_controlled() || n.get_owned_province_count() == 0)
-			return;
-
-		float base_income = economy::estimate_daily_income(state, n) + n.get_stockpiles(economy::money) / 365.f;
-
-		// they don't have to add up to 1.f
-		// the reason they are there is to slow down AI spendings,
-		// make them more or less balanced
-		// and stabilize economy faster
-		// not to allow it to hoard money
-
-		float land_budget_ratio					= 0.15f;
-		float sea_budget_ratio					= 0.05f;
-		float education_budget_ratio			= 0.30f;
-		float investments_budget_ratio			= 0.05f;
-		float soldiers_budget_ratio				= 0.30f;
-		float construction_budget_ratio			= 0.45f;
-		float overseas_maintenance_budget_ratio = 0.10f;
-
-		if(n.get_is_at_war()) {
-			land_budget_ratio *= 1.75f;
-			sea_budget_ratio *= 1.75f;
-			education_budget_ratio *= 0.15f;
-			overseas_maintenance_budget_ratio *= 0.15f;
-			//n.set_land_spending(int8_t(100));
-			//n.set_naval_spending(int8_t(100));
-		} else if(n.get_ai_is_threatened()) {
-			land_budget_ratio *= 1.25f;
-			sea_budget_ratio *= 1.25f;
-			education_budget_ratio *= 0.75f;
-			overseas_maintenance_budget_ratio *= 0.75f;
-			//n.set_land_spending(int8_t(50));
-			//n.set_naval_spending(int8_t(50));
-		} else {
-			//n.set_land_spending(int8_t(25));
-			//n.set_naval_spending(int8_t(25));
-		}
-		float land_budget = land_budget_ratio * base_income;
-		float naval_budget = sea_budget_ratio * base_income;
-		float education_budget = education_budget_ratio * base_income;
-		float construction_budget = construction_budget_ratio * base_income;
-		float soldiers_budget = soldiers_budget_ratio * base_income;
-		float overseas_budget = overseas_maintenance_budget_ratio * base_income;
-
-		float ratio_land = 100.f * land_budget / (1.f + economy::estimate_land_spending(state, n));
-		float ratio_naval = 100.f * naval_budget / (1.f + economy::estimate_naval_spending(state, n));
-
-		ratio_land = std::clamp(ratio_land, 0.f, 100.f);
-		ratio_naval = std::clamp(ratio_naval, 0.f, 100.f);
-		// Reduce spending at peace
-		if(!state.world.nation_get_is_at_war(n) && future_rebels_in_nation(state, n) == 0) {
-			ratio_land /= 10.f;
-			ratio_naval /= 10.f;
-		}
-		n.set_land_spending(int8_t(ratio_land));
-		n.set_naval_spending(int8_t(ratio_naval));
-
-		n.set_construction_spending(50);
-		n.set_administrative_spending(25);
-		
-		float max_education_budget = 1.f + economy::estimate_pop_payouts_by_income_type(state, n, culture::income_type::education);
-		float max_soldiers_budget = 1.f + economy::estimate_pop_payouts_by_income_type(state, n, culture::income_type::military);
-		float max_overseas_budget = 1.f + economy::estimate_overseas_penalty_spending(state, n);
-
-		// solving x^2 * max = desired
-		float ratio_education = 100.f * math::sqrt(education_budget / max_education_budget);
-		ratio_education = std::clamp(ratio_education, 0.f, 100.f);
-		n.set_education_spending(int8_t(ratio_education));		
-
-		// If State can build factories - why subsidize capitalists
-		auto rules = n.get_combined_issue_rules();
-		if(n.get_is_civilized() && (rules & issue_rule::build_factory) == 0) {
-			n.set_domestic_investment_spending(int8_t(investments_budget_ratio * 100.f));
-		} else {
-			n.set_domestic_investment_spending(int8_t(0));
-		}
-				
-		float soldiers_max_ratio = 100.f * math::sqrt(soldiers_budget / max_soldiers_budget);
-		soldiers_max_ratio = std::clamp(soldiers_max_ratio, 0.f, 100.f);
-
-		float overseas_max_ratio = std::clamp(100.f * overseas_budget / max_overseas_budget, 0.f, 100.f);
-
-		n.set_tariffs_import(int8_t(10));
-		n.set_tariffs_export(int8_t(10));
-
-		float poor_militancy = (state.world.nation_get_demographics(n, demographics::poor_militancy) / std::max(1.0f, state.world.nation_get_demographics(n, demographics::poor_total))) / 10.f;
-		float mid_militancy = (state.world.nation_get_demographics(n, demographics::middle_militancy) / std::max(1.0f, state.world.nation_get_demographics(n, demographics::middle_total))) / 10.f;
-		float rich_militancy = (state.world.nation_get_demographics(n, demographics::rich_militancy) / std::max(1.0f, state.world.nation_get_demographics(n, demographics::rich_total))) / 10.f;
-
-		// don't tax pops too much
-		if((rules & issue_rule::expand_factory) != 0 || (rules & issue_rule::build_factory) != 0) {
-			// Non-lf prioritize poor people
-			int max_poor_tax = int(10.f + 25.f * (1.f - poor_militancy));
-			int max_mid_tax = int(10.f + 35.f * (1.f - mid_militancy));
-			int max_rich_tax = int(10.f + 50.f * (1.f - rich_militancy));
-			int max_social = int(100.f * poor_militancy);
-
-			// enough tax?
-			bool enough_tax = true;
-			if(ratio_education < 50.f) {
-				enough_tax = false;
-				n.set_poor_tax(int8_t(std::clamp(n.get_poor_tax() + 2, 10, std::max(10, max_poor_tax))));
-				n.set_middle_tax(int8_t(std::clamp(n.get_middle_tax() + 3, 10, std::max(10, max_mid_tax))));
-				n.set_rich_tax(int8_t(std::clamp(n.get_rich_tax() + 5, 10, std::max(10, max_rich_tax))));
-			}
-
-			if(n.get_spending_level() < 1.0f || n.get_last_treasury() >= n.get_stockpiles(economy::money) || ratio_education < 50.f) { // losing money
-				if(!n.get_ai_is_threatened()) {
-					n.set_military_spending(int8_t(std::max(50, n.get_military_spending() - 5)));
-				}
-				n.set_social_spending(1);
-
-				n.set_poor_tax(int8_t(std::clamp(n.get_poor_tax() + 2, 10, std::max(10, max_poor_tax))));
-				n.set_middle_tax(int8_t(std::clamp(n.get_middle_tax() + 3, 10, std::max(10, max_mid_tax))));
-				n.set_rich_tax(int8_t(std::clamp(n.get_rich_tax() + 5, 10, std::max(10, max_rich_tax))));
-			} else if(n.get_last_treasury() < n.get_stockpiles(economy::money)) { // gaining money
-				if(n.get_ai_is_threatened()) {
-					n.set_military_spending(int8_t(std::min(100, n.get_military_spending() + 10)));
-				} else {
-					n.set_military_spending(int8_t(std::min(75, n.get_military_spending() + 10)));
-				}
-				n.set_social_spending(3);
-
-				if(enough_tax) {
-					n.set_poor_tax(int8_t(std::clamp(n.get_poor_tax() - 2, 10, std::max(10, max_poor_tax))));
-					n.set_middle_tax(int8_t(std::clamp(n.get_middle_tax() - 3, 10, std::max(10, max_mid_tax))));
-					n.set_rich_tax(int8_t(std::clamp(n.get_rich_tax() - 5, 10, std::max(10, max_rich_tax))));
-				}
-			}
-		} else {
-			int max_poor_tax = int(10.f + 30.f * (1.f - poor_militancy));
-			int max_mid_tax = int(10.f + 30.f * (1.f - mid_militancy));
-			int max_rich_tax = int(10.f + 10.f * (1.f - rich_militancy));
-			int max_social = int(20.f * poor_militancy);
-
-			// enough tax?
-			bool enough_tax = true;
-			if(ratio_education < 50.f) {
-				enough_tax = false;
-				n.set_poor_tax(int8_t(std::clamp(n.get_poor_tax() + 5, 10, std::max(10, max_poor_tax))));
-				n.set_middle_tax(int8_t(std::clamp(n.get_middle_tax() + 3, 10, std::max(10, max_mid_tax))));
-				n.set_rich_tax(int8_t(std::clamp(n.get_rich_tax() + 2, 10, std::max(10, max_rich_tax))));
-			}
-
-			// Laissez faire prioritize tax free capitalists
-			if(n.get_spending_level() < 1.0f || n.get_last_treasury() >= n.get_stockpiles(economy::money) || ratio_education < 50.f) { // losing money
-				if(!n.get_ai_is_threatened()) {
-					n.set_military_spending(int8_t(std::max(50, n.get_military_spending() - 5)));
-				}
-				n.set_social_spending(int8_t(max_social / 2));
-
-				n.set_poor_tax(int8_t(std::clamp(n.get_poor_tax() + 5, 10, std::max(10, max_poor_tax))));
-				n.set_middle_tax(int8_t(std::clamp(n.get_middle_tax() + 3, 10, std::max(10, max_mid_tax))));
-				n.set_rich_tax(int8_t(std::clamp(n.get_rich_tax() + 2, 10, std::max(10, max_rich_tax))));
-			} else if(n.get_last_treasury() < n.get_stockpiles(economy::money)) { // gaining money
-				if(n.get_ai_is_threatened()) {
-					n.set_military_spending(int8_t(std::min(100, n.get_military_spending() + 10)));
-				} else {
-					n.set_military_spending(int8_t(std::min(75, n.get_military_spending() + 10)));
-				}
-				n.set_social_spending(int8_t(max_social));
-
-				if(enough_tax) {
-					n.set_poor_tax(int8_t(std::clamp(n.get_poor_tax() - 5, 10, std::max(10, max_poor_tax))));
-					n.set_middle_tax(int8_t(std::clamp(n.get_middle_tax() - 3, 10, std::max(10, max_mid_tax))));
-					n.set_rich_tax(int8_t(std::clamp(n.get_rich_tax() - 2, 10, std::max(10, max_rich_tax))));
-				}
-			}
-		}
-
-		n.set_military_spending(int8_t(std::min(int8_t(soldiers_max_ratio), n.get_military_spending())));
-		n.set_overseas_spending(int8_t(overseas_max_ratio));
-
-		economy::bound_budget_settings(state, n);
-	});
-}
-
 
 void remove_ai_data(sys::state& state, dcon::nation_id n) {
 	for(auto ar : state.world.nation_get_army_control(n)) {
@@ -6292,23 +2501,6 @@ void general_ai_unit_tick(sys::state& state) {
 	case 7:
 		move_gathered_attackers(state);
 		break;
-	}
-
-	auto d = v % 100;
-
-	if(d == 0) {
-		for(auto nation : state.world.in_nation) {
-			if(state.world.nation_get_is_player_controlled(nation)) {
-				continue;
-			}
-
-			if(nation.get_is_great_power()) {
-				nation.set_ai_strategy(ai_strategies::militant);
-			}
-			else {
-				nation.set_ai_strategy(ai_strategies::industrious);
-			}
-		}
 	}
 }
