+++ src/ai/ai_focuses.hpp
@@ -1,0 +1,11 @@
+#pragma once
+#include "dcon_generated.hpp"
+
+namespace sys {
+struct state;
+}
+
+namespace ai {
+void update_focuses(sys::state& state);
+void identify_focuses(sys::state& state);
+}
