+++ src/ai/ai_types.hpp
@@ -1,0 +1,7 @@
+#pragma once
+
+namespace ai {
+enum ai_strategies {
+	industrious, militant, technological
+};
+}
