+++ src/gamestate/diplomatic_messages.cpp
@@ -2,7 +2,8 @@
 #include "system_state.hpp"
 #include "commands.hpp"
 #include "nations.hpp"
-#include "ai.hpp"
+#include "ai_campaign.hpp"
+#include "ai_war.hpp"
 
 namespace diplomatic_message {
