+++ src/gui/topbar_subwindows/diplomacy_subwindows/gui_diplomacy_actions_window.hpp
@@ -2,7 +2,8 @@
 
 #include "gui_element_types.hpp"
 #include "military.hpp"
-#include "ai.hpp"
+#include "ai_campaign_values.hpp"
+#include "ai_alliances.hpp"
 
 namespace ui {
