+++ src/launcher/launcher_main.cpp
@@ -51,6 +51,13 @@
 #include "map_tooltip.cpp"
 #include "unit_tooltip.cpp"
 #include "ai.cpp"
+#include "ai_campaign.cpp"
+#include "ai_campaign_values.cpp"
+#include "ai_focuses.cpp"
+#include "ai_alliances.cpp"
+#include "ai_influence.cpp"
+#include "ai_economy.cpp"
+#include "ai_war.cpp"
 #include "fif_triggers.cpp"
 #include "map_modes.cpp"
 #include "platform_specific.cpp"
